% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_scroll_box}
\alias{add_scroll_box}
\title{Add scroll box}
\usage{
add_scroll_box(
  table_kbl,
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  ...
)
}
\arguments{
\item{table_kbl}{PARAM_DESCRIPTION}

\item{scroll_height_1L_chr}{Scroll height (a character vector of length one), Default: character(0)}

\item{scroll_width_1L_chr}{Scroll width (a character vector of length one), Default: character(0)}

\item{...}{Additional arguments}
}
\value{
Table (a kable)
}
\description{
add_scroll_box() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add scroll box. The function is called for its side effects and does not return a value.
}
\keyword{internal}
