% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{author}
\alias{author}
\title{Author and save files}
\usage{
author(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
author() is a method that authors and saves files.
}
