% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorClasses}
\alias{authorClasses}
\title{Author and document classes}
\usage{
authorClasses(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
authorClasses() is a method that authors and saves code files for creating and documenting classes in an R package directory.
}
