% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorData}
\alias{authorData}
\title{Author and document package datasets}
\usage{
authorData(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
authorData() is a method that authors and saves code files for creating and documenting datasets in an R package directory.
}
