% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorFunctions}
\alias{authorFunctions}
\title{AuthorFunctions}
\usage{
authorFunctions(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorFunctions() is a method that authors and saves files necessary for creating and documenting functions.
}
