% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorReport}
\alias{authorReport}
\title{Author and save a report}
\usage{
authorReport(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
authorReport() is a method that authors and saves a report.
}
