% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bind.R
\name{bind_tables_from_loc_files}
\alias{bind_tables_from_loc_files}
\title{Bind tables from local files}
\usage{
bind_tables_from_loc_files(
  paths_chr,
  force_numeric_1L_lgl = FALSE,
  force_tb_1L_lgl = FALSE,
  heading_rows_1L_int = 1L
)
}
\arguments{
\item{paths_chr}{Paths (a character vector)}

\item{force_numeric_1L_lgl}{Force numeric (a logical vector of length one), Default: FALSE}

\item{force_tb_1L_lgl}{Force tibble (a logical vector of length one), Default: FALSE}

\item{heading_rows_1L_int}{Heading rows (an integer vector of length one), Default: 1}
}
\value{
Table (an output object of multiple potential types)
}
\description{
bind_tables_from_loc_files() is a Bind function that binds two objects together to create a composite object. Specifically, this function implements an algorithm to bind tables from local files. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
