% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{depict}
\alias{depict}
\title{Depict features of an instance of a class by generating a plot}
\usage{
depict(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
depict() is a method that depicts features of an instance of a class by generating a plot.
}
