% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated function}
\description{
Import of \code{\link[lifecycle:deprecated]{lifecycle::deprecated()}} function as recommended in lifecycle package documentation.
}
\seealso{
\code{\link[lifecycle:deprecated]{lifecycle::deprecated()}}
}
\keyword{internal}
