% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{.data}
\alias{.data}
\title{Dot Data function}
\description{
Import of \code{\link[rlang:dot-data]{rlang::.data()}} function for use in dataset manipulation within functions.
}
\seealso{
\code{\link[rlang:dot-data]{rlang::.data()}}
}
\keyword{internal}
