% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{enhanceSlot}
\alias{enhanceSlot}
\title{Enhance an instance of a class by adding new data fields and values to an object stored in a slot of that instance}
\usage{
enhanceSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
enhanceSlot() is a method that enhances an instance of a class by adding new data fields and values to an object stored in a slot of that instance.
}
