% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_exhibitSlot.R
\name{exhibitSlot}
\alias{exhibitSlot}
\alias{exhibitSlot-Ready4Module}
\alias{exhibitSlot,Ready4Module-method}
\title{Apply the exhibit method to a model module slot}
\usage{
exhibitSlot(x, slot_nm_1L_chr, ...)

\S4method{exhibitSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
Either a model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no return value (when called purely for side effects).

Either a ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no return value (when called purely for side effects).
}
\description{
exhibitSlot() is a convenience method that applies the exhibit method to a specified slot a model module.

exhibitSlot method applied to Ready4Module
}
