% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_badges_lup}
\alias{get_badges_lup}
\title{Get badges lookup table}
\usage{
get_badges_lup(
  ends_with_1L_chr = "ready4_badges_lup.RDS",
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{ends_with_1L_chr}{Ends with (a character vector of length one), Default: 'ready4_badges_lup.RDS'}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
Ready4 badges (a lookup table)
}
\description{
get_badges_lup() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get badges lookup table. The function returns Ready4 badges (a lookup table).
}
\keyword{internal}
