% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_cls_extensions}
\alias{get_cls_extensions}
\title{Get class extensions}
\usage{
get_cls_extensions(
  pkg_extensions_tb,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  validate_1L_lgl = F
)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble)}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{validate_1L_lgl}{Validate (a logical vector of length one), Default: F}
}
\value{
Class extensions (a tibble)
}
\description{
get_cls_extensions() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get class extensions. Function argument pkg_extensions_tb specifies the where to look for the required object. The function returns Class extensions (a tibble).
}
\keyword{internal}
