% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_datasets_tb}
\alias{get_datasets_tb}
\title{Get datasets tibble}
\usage{
get_datasets_tb(
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  rds_fl_name_1L_chr = "datasets_tb"
)
}
\arguments{
\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{rds_fl_name_1L_chr}{Rds file name (a character vector of length one), Default: 'datasets_tb'}
}
\value{
Datasets (a tibble)
}
\description{
get_datasets_tb() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get datasets tibble. The function returns Datasets (a tibble).
}
\examples{
get_datasets_tb("ready4-dev/ready4")
}
\keyword{internal}
