% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_digits_from_text}
\alias{get_digits_from_text}
\title{Get digits from text}
\usage{
get_digits_from_text(text_1L_chr)
}
\arguments{
\item{text_1L_chr}{Text (a character vector of length one)}
}
\value{
Digits (a character vector)
}
\description{
get_digits_from_text() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get digits from text. Function argument text_1L_chr specifies the where to look for the required object. The function returns Digits (a character vector).
}
\keyword{internal}
