% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dv_fls_urls}
\alias{get_dv_fls_urls}
\title{Get dataverse files urls}
\usage{
get_dv_fls_urls(
  file_nms_chr,
  dv_ds_nm_1L_chr,
  dv_url_pfx_1L_chr = character(0),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = NULL
)
}
\arguments{
\item{file_nms_chr}{File names (a character vector)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one)}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}
}
\value{
Urls (a character vector)
}
\description{
get_dv_fls_urls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get dataverse files urls. The function returns Urls (a character vector).
}
\keyword{internal}
