% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dvs}
\alias{get_dvs}
\title{Get dataverses}
\usage{
get_dvs(
  dv_nm_1L_chr = "ready4",
  key_1L_chr = NULL,
  server_1L_chr = "dataverse.harvard.edu"
)
}
\arguments{
\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one), Default: 'ready4'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}
}
\value{
Dataverses (a tibble)
}
\description{
get_dvs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get dataverses. Function argument dv_nm_1L_chr specifies the where to look for the required object. The function returns Dataverses (a tibble).
}
\keyword{internal}
