% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_examples}
\alias{get_examples}
\title{Get examples}
\usage{
get_examples(vignettes_chr, term_1L_chr)
}
\arguments{
\item{vignettes_chr}{Vignettes (a character vector)}

\item{term_1L_chr}{Term (a character vector of length one)}
}
\value{
Examples (a character vector)
}
\description{
get_examples() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get examples. Function argument vignettes_chr specifies the where to look for the required object. The function returns Examples (a character vector).
}
\keyword{internal}
