% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fl_extension}
\alias{get_fl_extension}
\title{Get file extension}
\usage{
get_fl_extension(path_1L_chr)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}
}
\value{
Extension (a character vector of length one)
}
\description{
get_fl_extension() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get file extension. The function returns Extension (a character vector of length one).
}
\keyword{internal}
