% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fl_id_from_dv_ls}
\alias{get_fl_id_from_dv_ls}
\title{Get file identity from dataverse list}
\usage{
get_fl_id_from_dv_ls(ds_ls, fl_nm_1L_chr, nms_chr = NA_character_)
}
\arguments{
\item{ds_ls}{Dataset (a list)}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{nms_chr}{Names (a character vector), Default: 'NA'}
}
\value{
Identity (a character vector of length one)
}
\description{
get_fl_id_from_dv_ls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get file identity from dataverse list. The function returns Identity (a character vector of length one).
}
\keyword{internal}
