% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fl_nm_from_path}
\alias{get_fl_nm_from_path}
\title{Get file name from path}
\usage{
get_fl_nm_from_path(path_1L_chr)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}
}
\value{
File name (a character vector of length one)
}
\description{
get_fl_nm_from_path() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get file name from path. The function returns File name (a character vector of length one).
}
\keyword{internal}
