% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_gh_repos}
\alias{get_gh_repos}
\title{Get github repositories}
\usage{
get_gh_repos(org_1L_chr)
}
\arguments{
\item{org_1L_chr}{Organisation (a character vector of length one)}
}
\value{
Repositories (a character vector)
}
\description{
get_gh_repos() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get github repositories. The function returns Repositories (a character vector).
}
\keyword{internal}
