% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_libraries_ls}
\alias{get_libraries_ls}
\title{Get libraries list}
\usage{
get_libraries_ls(
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
Libraries (a list)
}
\description{
get_libraries_ls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get libraries list. The function returns Libraries (a list).
}
\keyword{internal}
