% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_libraries_tb}
\alias{get_libraries_tb}
\title{Get a table of ready4 libraries}
\usage{
get_libraries_tb(
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
Libraries (a tibble)
}
\description{
get_libraries_tb() retrieves a tabular summary of ready4 libraries that have been developed within a specified GitHub organisation.
}
\examples{
get_libraries_tb("ready4-dev/ready4")
}
