% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_manual_urls}
\alias{get_manual_urls}
\title{Get manual urls}
\usage{
get_manual_urls(
  pkg_nm_1L_chr = "ready4",
  pkg_url_1L_chr = "https://ready4-dev.github.io/ready4/index.html"
)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{pkg_url_1L_chr}{Package url (a character vector of length one), Default: 'https://ready4-dev.github.io/ready4/index.html'}
}
\value{
Urls (a character vector)
}
\description{
get_manual_urls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get manual urls. Function argument pkg_nm_1L_chr specifies the where to look for the required object. The function returns Urls (a character vector).
}
\keyword{internal}
