% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_methods}
\alias{get_methods}
\title{Get methods}
\usage{
get_methods(pkg_nm_1L_chr = "ready4", cls_nm_1L_chr = "Ready4Module")
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{cls_nm_1L_chr}{Class name (a character vector of length one), Default: 'Ready4Module'}
}
\value{
Methods (a character vector)
}
\description{
get_methods() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get methods. Function argument pkg_nm_1L_chr specifies the where to look for the required object. The function returns Methods (a character vector).
}
\examples{
get_methods()
}
