% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_modules_tb}
\alias{get_modules_tb}
\title{Get a table of ready4 model modules}
\usage{
get_modules_tb(
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
Modules (a tibble)
}
\description{
get_modules_tb() ingests 'modules_tb.RDS' (a table of ready4 model modules) from a specified GitHub repository release.
}
\examples{
get_modules_tb("ready4-dev/ready4")
}
