% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_mthd_titles}
\alias{get_mthd_titles}
\title{Get method titles}
\usage{
get_mthd_titles(
  mthd_nms_chr,
  pkg_nm_1L_chr = "ready4",
  path_1L_chr = character(0)
)
}
\arguments{
\item{mthd_nms_chr}{Method names (a character vector)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}
}
\value{
Method titles (a character vector)
}
\description{
get_mthd_titles() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get method titles. Function argument mthd_nms_chr specifies the where to look for the required object. The function returns Method titles (a character vector).
}
\keyword{internal}
