% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_r4_obj_slots}
\alias{get_r4_obj_slots}
\title{Get ready4 S4 object slots}
\usage{
get_r4_obj_slots(fn_name_1L_chr, package_1L_chr = "")
}
\arguments{
\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{package_1L_chr}{Package (a character vector of length one), Default: ''}
}
\value{
Slots (a character vector)
}
\description{
get_r4_obj_slots() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get ready4 s4 object slots. Function argument fn_name_1L_chr specifies the where to look for the required object. The function returns Slots (a character vector).
}
\keyword{internal}
