% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_rds_from_dv}
\alias{get_rds_from_dv}
\title{Get rds from dataverse}
\usage{
get_rds_from_dv(
  file_nm_1L_chr,
  dv_ds_nm_1L_chr = "https://doi.org/10.7910/DVN/RIQTKK",
  dv_url_pfx_1L_chr = character(0),
  key_1L_chr = NULL,
  not_chr_1L_lgl = NA,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{file_nm_1L_chr}{File name (a character vector of length one)}

\item{dv_ds_nm_1L_chr}{Dataverse dataset name (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/RIQTKK'}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: character(0)}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{not_chr_1L_lgl}{Not character vector (a logical vector of length one), Default: NA}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
R object (an output object of multiple potential types)
}
\description{
get_rds_from_dv() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get rds from dataverse. The function returns R object (an output object of multiple potential types).
}
\keyword{internal}
