% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_source_code_urls}
\alias{get_source_code_urls}
\title{Get source code urls}
\usage{
get_source_code_urls(
  pkg_nm_1L_chr = "ready4",
  pkg_url_1L_chr = "https://ready4-dev.github.io/ready4/index.html"
)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{pkg_url_1L_chr}{Package url (a character vector of length one), Default: 'https://ready4-dev.github.io/ready4/index.html'}
}
\value{
Urls (a character vector)
}
\description{
get_source_code_urls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get source code urls. The function returns Urls (a character vector).
}
\keyword{internal}
