% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_table_from_loc_file}
\alias{get_table_from_loc_file}
\title{Get table from local file}
\usage{
get_table_from_loc_file(
  path_1L_chr,
  force_numeric_1L_lgl = F,
  force_tb_1L_lgl = F,
  heading_rows_1L_int = 1L
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{force_numeric_1L_lgl}{Force numeric (a logical vector of length one), Default: F}

\item{force_tb_1L_lgl}{Force tibble (a logical vector of length one), Default: F}

\item{heading_rows_1L_int}{Heading rows (an integer vector of length one), Default: 1}
}
\value{
Table (an output object of multiple potential types)
}
\description{
get_table_from_loc_file() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get table from local file. Function argument path_1L_chr specifies the where to look for the required object. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
