% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ingest}
\alias{ingest}
\title{Ingest data}
\usage{
ingest(x, ...)
}
\arguments{
\item{x}{A ready4 model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A ready4 model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance).
}
\description{
ingest() is a method that ingests data saved in external files into a ready4 model module or submodule.
}
