% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_ingestSlot.R
\name{ingestSlot}
\alias{ingestSlot}
\alias{ingestSlot-Ready4Module}
\alias{ingestSlot,Ready4Module-method}
\title{Apply the ingest method to a model module slot}
\usage{
ingestSlot(x, slot_nm_1L_chr, ...)

\S4method{ingestSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method.

A ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method.
}
\description{
ingestSlot() is a convenience method that applies the ingest method to a specified slot of a model module.

ingestSlot method applied to Ready4Module
}
