% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{investigate}
\alias{investigate}
\title{Investigate a dataset}
\usage{
investigate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
investigate() is a method that applies an algorithm to solve an inverse problem to data contained in an object.
}
