% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_investigateSlot.R
\name{investigateSlot}
\alias{investigateSlot}
\alias{investigateSlot-Ready4Module}
\alias{investigateSlot,Ready4Module-method}
\title{Apply the investigate method to a Ready4Module slot}
\usage{
investigateSlot(x, slot_nm_1L_chr, ...)

\S4method{investigateSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
investigateSlot() is a convenience method that applies the investigate method to a specified slot of an object.

investigateSlot method applied to Ready4Module
}
