% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_additions_tb}
\alias{make_additions_tb}
\title{Make additions tibble}
\usage{
make_additions_tb(
  category_chr = character(0),
  library_chr = character(0),
  type_chr = character(0),
  url_stub_1L_chr = "https://ready4-dev.github.io/"
)
}
\arguments{
\item{category_chr}{Category (a character vector), Default: character(0)}

\item{library_chr}{Library (a character vector), Default: character(0)}

\item{type_chr}{Type (a character vector), Default: character(0)}

\item{url_stub_1L_chr}{Url stub (a character vector of length one), Default: 'https://ready4-dev.github.io/'}
}
\value{
Additions (a tibble)
}
\description{
make_additions_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make additions tibble. The function returns Additions (a tibble).
}
\keyword{internal}
