% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_code_releases_tbl}
\alias{make_code_releases_tbl}
\title{Make code releases table}
\usage{
make_code_releases_tbl(
  repo_type_1L_chr = "Framework",
  as_kbl_1L_lgl = T,
  brochure_repos_chr = character(0),
  exclude_chr = character(0),
  format_1L_chr = "\%d-\%b-\%Y",
  framework_repos_chr = character(0),
  model_repos_chr = character(0),
  program_repos_chr = character(0),
  org_1L_chr = "ready4-dev",
  repos_chr = character(0),
  subroutine_repos_chr = character(0),
  tidy_desc_1L_lgl = T,
  url_stub_1L_chr = "https://ready4-dev.github.io/",
  ...
)
}
\arguments{
\item{repo_type_1L_chr}{Repository type (a character vector of length one), Default: 'Framework'}

\item{as_kbl_1L_lgl}{As kable (a logical vector of length one), Default: T}

\item{brochure_repos_chr}{Brochure repositories (a character vector), Default: character(0)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{format_1L_chr}{Format (a character vector of length one), Default: '\%d-\%b-\%Y'}

\item{framework_repos_chr}{Framework repositories (a character vector), Default: character(0)}

\item{model_repos_chr}{Model repositories (a character vector), Default: character(0)}

\item{program_repos_chr}{Program repositories (a character vector), Default: character(0)}

\item{org_1L_chr}{Organisation (a character vector of length one), Default: 'ready4-dev'}

\item{repos_chr}{Repositories (a character vector), Default: character(0)}

\item{subroutine_repos_chr}{Subroutine repositories (a character vector), Default: character(0)}

\item{tidy_desc_1L_lgl}{Tidy description (a logical vector of length one), Default: T}

\item{url_stub_1L_chr}{Url stub (a character vector of length one), Default: 'https://ready4-dev.github.io/'}

\item{...}{Additional arguments}
}
\value{
Releases (an output object of multiple potential types)
}
\description{
make_code_releases_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make code releases table. The function returns Releases (an output object of multiple potential types).
}
\keyword{internal}
