% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ds_releases_tbl}
\alias{make_ds_releases_tbl}
\title{Make dataset releases table}
\usage{
make_ds_releases_tbl(
  ds_dois_chr,
  format_1L_chr = "\%d-\%b-\%Y",
  server_1L_chr = "dataverse.harvard.edu"
)
}
\arguments{
\item{ds_dois_chr}{Dataset digital object identifiers (a character vector)}

\item{format_1L_chr}{Format (a character vector of length one), Default: '\%d-\%b-\%Y'}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}
}
\description{
make_ds_releases_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataset releases table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
