% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dvs_tb}
\alias{make_dvs_tb}
\title{Make dataverses tibble}
\usage{
make_dvs_tb(
  dv_nm_1L_chr = "ready4",
  key_1L_chr = NULL,
  server_1L_chr = "dataverse.harvard.edu"
)
}
\arguments{
\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one), Default: 'ready4'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}
}
\value{
Dataverses (a tibble)
}
\description{
make_dvs_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataverses tibble. The function returns Dataverses (a tibble).
}
