% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_files_tb}
\alias{make_files_tb}
\title{Make files tibble}
\usage{
make_files_tb(paths_to_dirs_chr, recode_ls, inc_fl_types_chr = NA_character_)
}
\arguments{
\item{paths_to_dirs_chr}{Paths to directories (a character vector)}

\item{recode_ls}{Recode (a list)}

\item{inc_fl_types_chr}{Include file types (a character vector), Default: 'NA'}
}
\value{
Files (a tibble)
}
\description{
make_files_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make files tibble. The function returns Files (a tibble).
}
\keyword{internal}
