% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_defaults_ls}
\alias{make_fn_defaults_ls}
\title{Make function defaults list}
\usage{
make_fn_defaults_ls(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Function defaults (a list)
}
\description{
make_fn_defaults_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function defaults list. The function returns Function defaults (a list).
}
\keyword{internal}
