% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_imports}
\alias{make_gnrc_imports}
\title{Make generic imports}
\usage{
make_gnrc_imports()
}
\value{
Generic imports (a character vector)
}
\description{
make_gnrc_imports() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make generic imports. The function returns Generic imports (a character vector).
}
