% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_libraries_ls}
\alias{make_libraries_ls}
\title{Make libraries list}
\usage{
make_libraries_ls(
  additions_tb = make_additions_tb(),
  libraries_tb = NULL,
  ns_var_nm_1L_chr = "pt_ns_chr"
)
}
\arguments{
\item{additions_tb}{Additions (a tibble), Default: make_additions_tb()}

\item{libraries_tb}{Libraries (a tibble), Default: NULL}

\item{ns_var_nm_1L_chr}{Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'}
}
\value{
Libraries (a list)
}
\description{
make_libraries_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make libraries list. The function returns Libraries (a list).
}
\keyword{internal}
