% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_list_phrase}
\alias{make_list_phrase}
\title{Make list phrase}
\usage{
make_list_phrase(items_chr)
}
\arguments{
\item{items_chr}{Items (a character vector)}
}
\value{
List phrase (a character vector of length one)
}
\description{
make_list_phrase() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make list phrase. The function returns List phrase (a character vector of length one).
}
\keyword{internal}
