% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_local_path_to_dv_data}
\alias{make_local_path_to_dv_data}
\title{Make local path to dataverse data}
\usage{
make_local_path_to_dv_data(save_dir_path_1L_chr, fl_nm_1L_chr, save_fmt_1L_chr)
}
\arguments{
\item{save_dir_path_1L_chr}{Save directory path (a character vector of length one)}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{save_fmt_1L_chr}{Save format (a character vector of length one)}
}
\value{
Path (a character vector)
}
\description{
make_local_path_to_dv_data() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make local path to dataverse data. The function returns Path (a character vector).
}
\keyword{internal}
