% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_methods_tb}
\alias{make_methods_tb}
\title{Make methods tibble}
\usage{
make_methods_tb(
  packages_tb = NULL,
  exclude_mthds_for_chr = NA_character_,
  include_1L_chr = "modules",
  ns_var_nm_1L_chr = "pt_ns_chr",
  reference_var_nm_1L_chr = "Reference",
  return_1L_lgl = "all",
  url_stub_1L_chr = "https://ready4-dev.github.io/",
  vignette_var_nm_1L_chr = "Vignettes",
  vignette_url_var_nm_1L_chr = "Vignettes_URLs"
)
}
\arguments{
\item{packages_tb}{Packages (a tibble), Default: NULL}

\item{exclude_mthds_for_chr}{Exclude methods for (a character vector), Default: 'NA'}

\item{include_1L_chr}{Include (a character vector of length one), Default: 'modules'}

\item{ns_var_nm_1L_chr}{Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'}

\item{reference_var_nm_1L_chr}{Reference variable name (a character vector of length one), Default: 'Reference'}

\item{return_1L_lgl}{Return (a logical vector of length one), Default: 'all'}

\item{url_stub_1L_chr}{Url stub (a character vector of length one), Default: 'https://ready4-dev.github.io/'}

\item{vignette_var_nm_1L_chr}{Vignette variable name (a character vector of length one), Default: 'Vignettes'}

\item{vignette_url_var_nm_1L_chr}{Vignette url variable name (a character vector of length one), Default: 'Vignettes_URLs'}
}
\value{
Methods (a tibble)
}
\description{
make_methods_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make methods tibble. The function returns Methods (a tibble).
}
