% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_modules_pkgs_chr}
\alias{make_modules_pkgs_chr}
\title{Make modules packages character vector}
\usage{
make_modules_pkgs_chr(what_chr = "all")
}
\arguments{
\item{what_chr}{What (a character vector), Default: 'all'}
}
\value{
Modules packages (a character vector)
}
\description{
make_modules_pkgs_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make modules packages character vector. The function returns Modules packages (a character vector).
}
\keyword{internal}
