% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pkg_extensions_tb}
\alias{make_pkg_extensions_tb}
\title{Make package extensions tibble}
\usage{
make_pkg_extensions_tb()
}
\value{
Package extensions (a tibble)
}
\description{
make_pkg_extensions_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make package extensions tibble. The function returns Package extensions (a tibble).
}
\keyword{internal}
