% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_programs_tbl}
\alias{make_programs_tbl}
\title{Make programs table}
\usage{
make_programs_tbl(
  what_1L_chr = "Program",
  as_kbl_1L_lgl = F,
  exclude_chr = "dce_sa_cards",
  format_1L_chr = "\%d-\%b-\%Y",
  tidy_desc_1L_lgl = T,
  zenodo_1L_chr = "ready4",
  ...
)
}
\arguments{
\item{what_1L_chr}{What (a character vector of length one), Default: 'Program'}

\item{as_kbl_1L_lgl}{As kable (a logical vector of length one), Default: F}

\item{exclude_chr}{Exclude (a character vector), Default: 'dce_sa_cards'}

\item{format_1L_chr}{Format (a character vector of length one), Default: '\%d-\%b-\%Y'}

\item{tidy_desc_1L_lgl}{Tidy description (a logical vector of length one), Default: T}

\item{zenodo_1L_chr}{Zenodo (a character vector of length one), Default: 'ready4'}

\item{...}{Additional arguments}
}
\value{
Programs (an output object of multiple potential types)
}
\description{
make_programs_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make programs table. The function returns Programs (an output object of multiple potential types).
}
\keyword{internal}
