% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_prompt}
\alias{make_prompt}
\title{Make prompt}
\usage{
make_prompt(prompt_1L_chr, options_chr = NULL, force_from_opts_1L_chr = FALSE)
}
\arguments{
\item{prompt_1L_chr}{Prompt (a character vector of length one)}

\item{options_chr}{Options (a character vector), Default: NULL}

\item{force_from_opts_1L_chr}{Force from opts (a character vector of length one), Default: FALSE}
}
\value{
Response (a character vector of length one)
}
\description{
make_prompt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make prompt. The function returns Response (a character vector of length one).
}
\keyword{internal}
