% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_subroutine_repos}
\alias{make_subroutine_repos}
\title{Make subroutine repositories}
\usage{
make_subroutine_repos()
}
\value{
Subroutine repositories (a character vector)
}
\description{
make_subroutine_repos() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make subroutine repositories. The function returns Subroutine repositories (a character vector).
}
\keyword{internal}
