% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{manufacture}
\alias{manufacture}
\title{Manufacture a new object}
\usage{
manufacture(x, ...)
}
\arguments{
\item{x}{A ready4 model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
An object other than a ready4 model module (an instance of a class that inherits from Ready4Module).
}
\description{
manufacture() is a method that creates a new object (other than a ready4 model module).
}
