% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_manufactureSlot.R
\name{manufactureSlot}
\alias{manufactureSlot}
\alias{manufactureSlot-Ready4Module}
\alias{manufactureSlot,Ready4Module-method}
\title{Apply the manufacture method to a slot}
\usage{
manufactureSlot(x, slot_nm_1L_chr, ...)

\S4method{manufactureSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
manufactureSlot() is a convenience method that applies the manufacture method to a specified slot of an object.

manufactureSlot method applied to Ready4Module
}
