% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Metamorphose an object}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
metamorphose() is a method that transforms one ready4 object into a ready4 object of a different class.
}
