% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphoseSlot}
\alias{metamorphoseSlot}
\title{Metamorphose data stored in a slot of a class instance into an object of a different (non-child) class}
\usage{
metamorphoseSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
metamorphoseSlot() is a method that metamorphoses data stored in a slot of a class instance into an object of a different (non-child) class.
}
