% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{:=}
\alias{:=}
\title{Non standard evaluation assignment function}
\description{
Import of := function, implementing \code{\link[rlang:dyn-dots]{rlang::dyn-dots()}}. See rlang package documentation for detailed description.
}
\seealso{
\code{\link[rlang:dyn-dots]{rlang::dyn-dots()}}
}
\keyword{internal}
