% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\description{
Import of \code{\link[magrittr:pipe]{magrittr::\%>\%()}}. See magrittr package documentation for detailed description.
}
\seealso{
\code{\link[magrittr:pipe]{magrittr::\%>\%()}}
}
\keyword{internal}
