% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_data}
\alias{print_data}
\title{Print data}
\usage{
print_data(
  datasets_tb,
  by_dv_1L_lgl = F,
  root_1L_chr = "https://dataverse.harvard.edu/dataverse/",
  what_1L_chr = "all"
)
}
\arguments{
\item{datasets_tb}{Datasets (a tibble)}

\item{by_dv_1L_lgl}{By dataverse (a logical vector of length one), Default: F}

\item{root_1L_chr}{Root (a character vector of length one), Default: 'https://dataverse.harvard.edu/dataverse/'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}
}
\value{
Datasets (a kable)
}
\description{
print_data() is a Print function that prints output to console Specifically, this function implements an algorithm to print data. The function is called for its side effects and does not return a value.
}
\keyword{internal}
