% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_dss}
\alias{print_dss}
\title{Print datasets}
\usage{
print_dss(dvs_tb, what_1L_chr = "all")
}
\arguments{
\item{dvs_tb}{Dataverses (a tibble)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}
}
\value{
dss_kbl (An object)
}
\description{
print_dss() is a Print function that prints output to console Specifically, this function implements an algorithm to print datasets. The function is called for its side effects and does not return a value.
}
\keyword{internal}
