% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_dvs}
\alias{print_dvs}
\title{Print dataverses}
\usage{
print_dvs(
  dvs_tb,
  root_1L_chr = "https://dataverse.harvard.edu/dataverse/",
  what_1L_chr = "all"
)
}
\arguments{
\item{dvs_tb}{Dataverses (a tibble)}

\item{root_1L_chr}{Root (a character vector of length one), Default: 'https://dataverse.harvard.edu/dataverse/'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}
}
\value{
Dataverses (a kable)
}
\description{
print_dvs() is a Print function that prints output to console Specifically, this function implements an algorithm to print dataverses. The function is called for its side effects and does not return a value.
}
