% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_methods}
\alias{print_methods}
\title{Print methods}
\usage{
print_methods(
  methods_tb = NULL,
  exclude_mthds_for_chr = NA_character_,
  methods_chr = NULL,
  path_1L_chr = character(0),
  return_1L_chr = "all"
)
}
\arguments{
\item{methods_tb}{Methods (a tibble), Default: NULL}

\item{exclude_mthds_for_chr}{Exclude methods for (a character vector), Default: 'NA'}

\item{methods_chr}{Methods (a character vector), Default: NULL}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
methods_kbl (An object)
}
\description{
print_methods() is a Print function that prints output to console Specifically, this function implements an algorithm to print methods. The function is called for its side effects and does not return a value.
}
