% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_pkg_extensions}
\alias{print_pkg_extensions}
\title{Print package extensions}
\usage{
print_pkg_extensions(pkg_extensions_tb = NULL)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble), Default: NULL}
}
\value{
pkg_extensions_kbl (An object)
}
\description{
print_pkg_extensions() is a Print function that prints output to console Specifically, this function implements an algorithm to print package extensions. The function is called for its side effects and does not return a value.
}
\keyword{internal}
