% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procure}
\alias{procure}
\title{Procure items from a dataset}
\usage{
procure(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
procure() is a method that retrieves data contained within a model module or sub-module instance. A "getter" method.
}
