% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procureSlot}
\alias{procureSlot}
\title{Procure data contained in a slot of a class instance}
\usage{
procureSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
procureSlot() is a method that procures data contained in a slot of an instance of a class.
}
