% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ratify}
\alias{ratify}
\title{Check whether an instance of a class conforms to required criteria, potentially modifying that instance to ensure that it is valid.}
\usage{
ratify(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
ratify() is a method that checks whether an instance of a class conforms to required criteria, potentially modifying that instance to ensure that it is valid.
}
