% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_ratifySlot.R
\name{ratifySlot}
\alias{ratifySlot}
\alias{ratifySlot-Ready4Module}
\alias{ratifySlot,Ready4Module-method}
\title{Apply the ratify method to a Ready4Module slot}
\usage{
ratifySlot(x, slot_nm_1L_chr, ...)

\S4method{ratifySlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
ratifySlot() is a method that validates that data in a slot of an S4 object conforms to specified criteria, potentially modifying that data so that these criteria are met.

ratifySlot method applied to Ready4Module
}
