% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4.R, R/ready4-package.R
\docType{package}
\name{ready4-package}
\alias{ready4}
\alias{ready4-package}
\title{ready4: Implement Modular Health Economic Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A prototype software framework to support collaborative and modular implementation of health economic models. To help make health economic models more Transparent, Reusable And Updatable (TRU), 'ready4' provides a simple programming syntax, a template model module, and tools for documenting model implementations. These foundational elements of the 'ready4' software framework are extended by other R packages. For detailed documentation about how to use 'ready4' and its extensions visit \url{https://www.ready4-dev.com/}. For a background to and rationale for transparent, reusable and updatable computational health economic models read the manuscript \href{https://arxiv.org/abs/2310.14138}{arXiv:2310.14138}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A prototype software framework to support collaborative and modular implementation of health economic models. To help make health economic models more Transparent, Reusable And Updatable (TRU), 'ready4' provides a simple programming syntax, a template model module, and tools for documenting model implementations. These foundational elements of the 'ready4' software framework are extended by other R packages. For detailed documentation about how to use 'ready4' and its extensions visit \url{https://www.ready4-dev.com/}. For a background to and rationale for transparent, reusable and updatable computational health economic models read the manuscript \href{https://arxiv.org/abs/2310.14138}{arXiv:2310.14138}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
