% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Launch.R
\docType{class}
\name{Ready4Launch-class}
\alias{Ready4Launch-class}
\alias{Ready4Launch}
\title{Ready4Launch}
\description{
A virtual S4 class that is the foundation of the ready4 representational system.
}
