% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Module.R
\docType{class}
\name{Ready4Module-class}
\alias{Ready4Module-class}
\alias{Ready4Module}
\title{Ready4Module}
\description{
A module of the ready4 representational system.
}
\section{Slots}{

\describe{
\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

