% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Private.R
\docType{class}
\name{Ready4Private-class}
\alias{Ready4Private-class}
\alias{Ready4Private}
\title{Ready4Private}
\description{
A virtual S4 class used to signify that a class extending the ready4 representational system contains data not intended for public dissemination.
}
