% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Project.R
\docType{class}
\name{Ready4Project-class}
\alias{Ready4Project-class}
\alias{Ready4Project}
\title{Ready4Project}
\description{
A module of the ready4 framework that contains details of a modelling project.
}
\section{Slots}{

\describe{
\item{\code{ds_tb}}{Dataset (a tibble)}

\item{\code{dictionary_r3}}{Dictionary (a ready4 submodule)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

