% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Public.R
\docType{class}
\name{Ready4Public-class}
\alias{Ready4Public-class}
\alias{Ready4Public}
\title{Ready4Public}
\description{
A virtual S4 class used to signify that a class extending the ready4 representational system is exclusively comprised of data appropriate for public dissemination.
}
