% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{reckon}
\alias{reckon}
\title{Reckon a value by performing a calculation using data contained in an instance of a class}
\usage{
reckon(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
reckon() is a method that reckons a value by performing a calculation using data contained in an instance of a class.
}
