% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_lbls_from_df}
\alias{remove_lbls_from_df}
\title{Remove labels from dataframe}
\usage{
remove_lbls_from_df(data_df)
}
\arguments{
\item{data_df}{Data (a data.frame)}
}
\value{
Unlabelled data (a data.frame)
}
\description{
remove_lbls_from_df() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove labels from dataframe. The function returns Unlabelled data (a data.frame).
}
\keyword{internal}
