% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_renewSlot.R
\name{renewSlot}
\alias{renewSlot}
\alias{renewSlot-Ready4Module}
\alias{renewSlot,Ready4Module-method}
\title{Renew (update) an element of an object}
\usage{
renewSlot(x, ...)

\S4method{renewSlot}{Ready4Module}(x, slot_nm_1L_chr, new_val_xx)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{new_val_xx}{New value (slot dependent object type)}
}
\description{
renewSlot() is a "setter" method that updates a slot of a S4 object with new data.

renewSlot method applied to Ready4Module
}
