% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{report}
\alias{report}
\title{Author a report.}
\usage{
report(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
report() is a method that authors a report.
}
