% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rowbind.R
\name{rowbind_tbs_in_r4_obj}
\alias{rowbind_tbs_in_r4_obj}
\title{rowbind tibbles in ready4 S4 object}
\usage{
rowbind_tbs_in_r4_obj(tbs_r4, slot_nm_1L_chr, second_tbs_r4, r4_name_1L_chr)
}
\arguments{
\item{tbs_r4}{Tibbles (a ready4 S4)}

\item{slot_nm_1L_chr}{Slot name (a character vector of length one)}

\item{second_tbs_r4}{Second tibbles (a ready4 S4)}

\item{r4_name_1L_chr}{Ready4 S4 name (a character vector of length one)}
}
\value{
Tibbles (a ready4 S4)
}
\description{
rowbind_tbs_in_r4_obj() is a Rowbind function that performs custom rowbind operations on table objects. Specifically, this function implements an algorithm to rowbind tibbles in ready4 s4 object. The function returns Tibbles (a ready4 S4).
}
\keyword{internal}
