% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{share}
\alias{share}
\title{Share (publish) open data}
\usage{
share(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
share() is a method that uploads data to an online repository. If requested, the method will also publish the updated repository.
}
