% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_shareSlot.R
\name{shareSlot}
\alias{shareSlot}
\alias{shareSlot-Ready4Module}
\alias{shareSlot,Ready4Module-method}
\title{Share data contained in a slot of a class instance via an online repository}
\usage{
shareSlot(x, ...)

\S4method{shareSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
shareSlot() is a method that shares data contained in a slot of a class instance via an online repository.

shareSlot method applied to Ready4Module
}
