% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_chr_to_num}
\alias{transform_chr_to_num}
\title{Transform character vector to numeric}
\usage{
transform_chr_to_num(digits_chr)
}
\arguments{
\item{digits_chr}{Digits (a character vector)}
}
\value{
Digits (an output object of multiple potential types)
}
\description{
transform_chr_to_num() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform character vector to numeric. The function returns Digits (an output object of multiple potential types).
}
\keyword{internal}
