% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_cls_type_ls}
\alias{transform_cls_type_ls}
\title{Transform class type list}
\usage{
transform_cls_type_ls(cls_type_ls)
}
\arguments{
\item{cls_type_ls}{Class type (a list)}
}
\value{
Tfmd class type (a list)
}
\description{
transform_cls_type_ls() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform class type list. The function returns Tfmd class type (a list).
}
\keyword{internal}
