% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_libraries_ls}
\alias{update_libraries_ls}
\title{Update libraries list}
\usage{
update_libraries_ls(
  libraries_ls = NULL,
  additions_tb = make_additions_tb(),
  keep_chr = "all"
)
}
\arguments{
\item{libraries_ls}{Libraries (a list), Default: NULL}

\item{additions_tb}{Additions (a tibble), Default: make_additions_tb()}

\item{keep_chr}{Keep (a character vector), Default: 'all'}
}
\value{
Libraries (a list)
}
\description{
update_libraries_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update libraries list. The function returns Libraries (a list).
}
\keyword{internal}
