% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_pt_fn_args_ls}
\alias{update_pt_fn_args_ls}
\title{Update prototype function arguments list}
\usage{
update_pt_fn_args_ls(args_ls)
}
\arguments{
\item{args_ls}{Arguments (a list)}
}
\value{
Updated arguments (a list)
}
\description{
update_pt_fn_args_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update prototype function arguments list. The function returns Updated arguments (a list).
}
\keyword{internal}
