% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_tb_r3}
\alias{update_tb_r3}
\title{Update tibble ready4 S3}
\usage{
update_tb_r3(
  tb_r3,
  filter_cdn_1L_chr = NA_character_,
  slice_idxs_int = NA_integer_
)
}
\arguments{
\item{tb_r3}{Tibble ready4 S3 (a ready4 S3 extension of tibble)}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{slice_idxs_int}{Slice indices (an integer vector), Default: NA}
}
\value{
Tibble ready4 S3 (a ready4 S3 extension of tibble)
}
\description{
update_tb_r3() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update tibble ready4 s3. Function argument tb_r3 specifies the object to be updated. Argument filter_cdn_1L_chr provides the object to be updated. The function returns Tibble ready4 S3 (a ready4 S3 extension of tibble).
}
\keyword{internal}
