% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_citation_cff}
\alias{write_citation_cff}
\title{Write citation cff}
\usage{
write_citation_cff(
  pkg_desc_ls,
  citation_chr,
  consent_1L_chr = "",
  publisher_1L_chr = "Zenodo"
)
}
\arguments{
\item{pkg_desc_ls}{Package description (a list)}

\item{citation_chr}{Citation (a character vector)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{publisher_1L_chr}{Publisher (a character vector of length one), Default: 'Zenodo'}
}
\description{
write_citation_cff() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write citation cff. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
