% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_conditional_tags}
\alias{write_conditional_tags}
\title{Write conditional tags}
\usage{
write_conditional_tags(
  pkgs_chr,
  path_to_pkg_root_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  where_1L_chr = character(0)
)
}
\arguments{
\item{pkgs_chr}{Packages (a character vector)}

\item{path_to_pkg_root_1L_chr}{Path to package root (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{where_1L_chr}{Where (a character vector of length one), Default: character(0)}
}
\value{
No return value, called for side effects.
}
\description{
write_conditional_tags() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write conditional tags. The function is called for its side effects and does not return a value.
}
\seealso{
\code{\link[usethis:use_package]{usethis::use_package()}}
}
\keyword{internal}
