% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dv_fl_to_loc}
\alias{write_dv_fl_to_loc}
\title{write dataverse file to local}
\usage{
write_dv_fl_to_loc(
  ds_ui_1L_chr,
  fl_nm_1L_chr = NA_character_,
  fl_id_1L_int = NA_integer_,
  repo_fl_fmt_1L_chr,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  save_type_1L_chr = "original",
  dest_path_1L_chr,
  consent_1L_chr = ""
)
}
\arguments{
\item{ds_ui_1L_chr}{Dataset user interface (a character vector of length one)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'NA'}

\item{fl_id_1L_int}{File identity (an integer vector of length one), Default: NA}

\item{repo_fl_fmt_1L_chr}{Repository file format (a character vector of length one)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{save_type_1L_chr}{Save type (a character vector of length one), Default: 'original'}

\item{dest_path_1L_chr}{Destination path (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}
}
\description{
write_dv_fl_to_loc() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write dataverse file to local. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
