% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_env_objs_to_dv}
\alias{write_env_objs_to_dv}
\title{Write environment objects to dataverse}
\usage{
write_env_objs_to_dv(
  env_objects_ls,
  descriptions_chr,
  ds_url_1L_chr,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  publish_dv_1L_lgl = F,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{env_objects_ls}{Environment objects (a list)}

\item{descriptions_chr}{Descriptions (a character vector)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: F}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
File identities (an integer vector)
}
\description{
write_env_objs_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write environment objects to dataverse. The function returns File identities (an integer vector).
}
\keyword{internal}
