% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_examples}
\alias{write_examples}
\title{Write examples}
\usage{
write_examples(
  path_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  type_1L_chr = "fn"
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'fn'}
}
\value{
No return value, called for side effects.
}
\description{
write_examples() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write examples. The function is called for its side effects and does not return a value.
}
\seealso{
\code{\link[devtools:document]{devtools::document()}}
}
\keyword{internal}
