% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_extra_pkgs_to_actions}
\alias{write_extra_pkgs_to_actions}
\title{Write extra packages to actions}
\usage{
write_extra_pkgs_to_actions(path_to_dir_1L_chr = ".github/workflows")
}
\arguments{
\item{path_to_dir_1L_chr}{Path to directory (a character vector of length one), Default: '.github/workflows'}
}
\description{
write_extra_pkgs_to_actions() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write extra packages to actions. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
