% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_from_dv}
\alias{write_fls_from_dv}
\title{Write files from dataverse}
\usage{
write_fls_from_dv(
  files_tb,
  fl_ids_int,
  ds_url_1L_chr,
  local_dv_dir_1L_chr,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  consent_1L_chr = ""
)
}
\arguments{
\item{files_tb}{Files (a tibble)}

\item{fl_ids_int}{File identities (an integer vector)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{local_dv_dir_1L_chr}{Local dataverse directory (a character vector of length one)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}
}
\description{
write_fls_from_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files from dataverse. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
