% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_to_dv}
\alias{write_fls_to_dv}
\title{write files to dataverse}
\usage{
write_fls_to_dv(
  file_paths_chr,
  descriptions_chr = NULL,
  ds_url_1L_chr,
  ds_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{file_paths_chr}{File paths (a character vector)}

\item{descriptions_chr}{Descriptions (a character vector), Default: NULL}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{ds_ls}{Dataset (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Identities (an integer vector)
}
\description{
write_fls_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files to dataverse. The function returns Identities (an integer vector).
}
\keyword{internal}
