% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_new_files}
\alias{write_new_files}
\title{Write new files}
\usage{
write_new_files(
  paths_chr,
  custom_write_ls = NULL,
  fl_nm_1L_chr = NULL,
  source_paths_ls = NULL,
  text_ls = NULL
)
}
\arguments{
\item{paths_chr}{Paths (a character vector)}

\item{custom_write_ls}{Custom write (a list), Default: NULL}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: NULL}

\item{source_paths_ls}{Source paths (a list), Default: NULL}

\item{text_ls}{Text (a list), Default: NULL}
}
\description{
write_new_files() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
