% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_prj_outp_dirs}
\alias{write_prj_outp_dirs}
\title{Write project output directories}
\usage{
write_prj_outp_dirs(
  prj_dirs_chr,
  output_data_dir_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  paths_ls = NULL
)
}
\arguments{
\item{prj_dirs_chr}{Project directories (a character vector)}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{paths_ls}{Paths (a list), Default: NULL}
}
\value{
New paths (a list)
}
\description{
write_prj_outp_dirs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write project output directories. The function returns New paths (a list).
}
\keyword{internal}
