% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_delete_dirs}
\alias{write_to_delete_dirs}
\title{write to delete directories}
\usage{
write_to_delete_dirs(dir_paths_chr)
}
\arguments{
\item{dir_paths_chr}{Directory paths (a character vector)}
}
\description{
write_to_delete_dirs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to delete directories. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
