% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_publish_dv_ds}
\alias{write_to_publish_dv_ds}
\title{Write to publish dataverse dataset}
\usage{
write_to_publish_dv_ds(dv_ds_1L_chr)
}
\arguments{
\item{dv_ds_1L_chr}{Dataverse dataset (a character vector of length one)}
}
\description{
write_to_publish_dv_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to publish dataverse dataset. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
