% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_with_consent}
\alias{write_with_consent}
\title{Write with consent}
\usage{
write_with_consent(
  consented_fn,
  prompt_1L_chr,
  consent_1L_chr = "",
  consented_args_ls = NULL,
  consent_indcs_int = 1L,
  consented_msg_1L_chr = character(0),
  declined_args_ls = NULL,
  declined_fn = NULL,
  declined_msg_1L_chr = "No files have been written.",
  force_from_opts_1L_chr = TRUE,
  options_chr = c("Y", "N"),
  return_1L_lgl = FALSE
)
}
\arguments{
\item{consented_fn}{Consented (a function)}

\item{prompt_1L_chr}{Prompt (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consented_args_ls}{Consented arguments (a list), Default: NULL}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{consented_msg_1L_chr}{Consented message (a character vector of length one), Default: character(0)}

\item{declined_args_ls}{Declined arguments (a list), Default: NULL}

\item{declined_fn}{Declined (a function), Default: NULL}

\item{declined_msg_1L_chr}{Declined message (a character vector of length one), Default: 'No files have been written.'}

\item{force_from_opts_1L_chr}{Force from opts (a character vector of length one), Default: TRUE}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{return_1L_lgl}{Return (a logical vector of length one), Default: FALSE}
}
\value{
Object (an output object of multiple potential types)
}
\description{
write_with_consent() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write with consent. The function returns Object (an output object of multiple potential types).
}
\keyword{internal}
