% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_lups}
\alias{add_lups}
\title{Add lookup tables}
\usage{
add_lups(
  template_lup,
  new_lup,
  key_var_nm_1L_chr,
  priority_lup_for_dupls_1L_chr = "template"
)
}
\arguments{
\item{template_lup}{Template (a lookup table)}

\item{new_lup}{New (a lookup table)}

\item{key_var_nm_1L_chr}{Key variable name (a character vector of length one)}

\item{priority_lup_for_dupls_1L_chr}{Priority lookup table for duplicates (a character vector of length one), Default: 'template'}
}
\value{
Combined (lookup tables)
}
\description{
add_lups() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add lookup tables. The function is called for its side effects and does not return a value.
}
\seealso{
\code{\link[Hmisc:label]{Hmisc::label()}}
}
\keyword{internal}
