% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_rows_from_fn_args}
\alias{add_rows_from_fn_args}
\title{Add rows from function arguments}
\usage{
add_rows_from_fn_args(tbl_r3, fn, fn_env_ls)
}
\arguments{
\item{tbl_r3}{Table (a ready4 submodule)}

\item{fn}{Function (a function)}

\item{fn_env_ls}{Function (a list of environments)}
}
\value{
Table (a ready4 submodule)
}
\description{
add_rows_from_fn_args() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add rows from function arguments. The function returns Table (a ready4 submodule).
}
\seealso{
\code{\link[Hmisc:label]{Hmisc::label()}}
}
\keyword{internal}
