% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{author}
\alias{author}
\title{Author}
\usage{
author(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
author() is a method that writes files to local or remote locations.
}
