% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{author}
\alias{author}
\title{Author}
\usage{
author(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
author() is an Author function that writes files to local or remote locations. Specifically, this function implements an algorithm to author. The function is called for its side effects and does not return a value.
}
