% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{author}
\alias{author}
\title{Author and save files}
\usage{
author(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
Either a model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance) of the same class as that supplied to the method or no return value (when called for side-effects only).
}
\description{
author() is a method that authors and saves files.
}
