% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorClasses}
\alias{authorClasses}
\title{AuthorClasses}
\usage{
authorClasses(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorClasses() is a method that authors and saves files necessary for creating and documenting classes.
}
