% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorClasses}
\alias{authorClasses}
\title{AuthorClasses}
\usage{
authorClasses(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorClasses() is an AuthorClasses function that authors and saves files necessary for creating and documenting classes. Specifically, this function implements an algorithm to authorclasses. The function is called for its side effects and does not return a value.
}
