% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorData}
\alias{authorData}
\title{AuthorData}
\usage{
authorData(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorData() is a method that authors and saves files necessary for creating and documenting datasets.
}
