% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorData}
\alias{authorData}
\title{AuthorData}
\usage{
authorData(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorData() is an AuthorData function that authors and saves files necessary for creating and documenting datasets. Specifically, this function implements an algorithm to authordata. The function is called for its side effects and does not return a value.
}
