% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{authorFunctions}
\alias{authorFunctions}
\title{AuthorFunctions}
\usage{
authorFunctions(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
authorFunctions() is an AuthorFunctions function that authors and saves files necessary for creating and documenting functions. Specifically, this function implements an algorithm to authorfunctions. The function is called for its side effects and does not return a value.
}
