% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_authorSlot.R
\name{authorSlot}
\alias{authorSlot}
\alias{authorSlot-Ready4Module}
\alias{authorSlot,Ready4Module-method}
\title{Apply the author method to a model module slot}
\usage{
authorSlot(x, slot_nm_1L_chr, ...)

\S4method{authorSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
Either a model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called for side effects only).

Either a ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called for side effects only).
}
\description{
authorSlot() is a convenience method that applies the author method to a specified slot of a model module.

authorSlot method applied to Ready4Module
}
