% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterize}
\alias{characterize}
\title{Characterize data by generating (tabular) descriptive statistics}
\usage{
characterize(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
characterize() is a method that generates descriptive tabular summaries about a dataset.
}
