% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterize}
\alias{characterize}
\title{Generate descriptive information about a dataset.}
\usage{
characterize(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
characterize() is a method that generates descriptive information about a dataset.
}
