% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterize}
\alias{characterize}
\title{Characterize the features of a ready4 framework module (or sub-module)}
\usage{
characterize(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
characterize() is a method that generates descriptive information about the features of a ready4 framework module (or sub-module).
}
