% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterize}
\alias{characterize}
\title{Characterize}
\usage{
characterize(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
characterize() is a Characterize function that generates descriptive information about a dataset. Specifically, this function implements an algorithm to characterize. The function is called for its side effects and does not return a value.
}
