% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterize}
\alias{characterize}
\title{Characterize model module data by generating (tabular) descriptive statistics}
\usage{
characterize(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A data.frame, tibble or other table based class.
}
\description{
characterize() is a method that generates descriptive tabular summaries about data contained in a model module.
}
