% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{characterizeSlot}
\alias{characterizeSlot}
\title{Characterize a an object stored in a slot of a class instance by generating descriptive information}
\usage{
characterizeSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
characterizeSlot() is a method that characterizes a an object stored in a slot of a class instance by generating descriptive information.
}
