% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{depict}
\alias{depict}
\title{Depict (plot) features of a dataset}
\usage{
depict(x, ...)
}
\arguments{
\item{x}{A ready4 model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A ggplot, gg or other plot type class.
}
\description{
depict() is a method that plots features of a ready4 framework module (or sub-module).
}
