% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{depict}
\alias{depict}
\title{Depict (plot) features of a dataset}
\usage{
depict(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
depict() is a method that plots features of a ready4 framework module (or sub-module).
}
