% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_depictSlot.R
\name{depictSlot}
\alias{depictSlot}
\alias{depictSlot-Ready4Module}
\alias{depictSlot,Ready4Module-method}
\title{Apply the depict method to a model module slot}
\usage{
depictSlot(x, slot_nm_1L_chr, ...)

\S4method{depictSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
Either a model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called for side effects only).

Either a ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called for side effects only).
}
\description{
depictSlot() is a convenience method that applies the depict method to a specified slot of a model module.

depictSlot method applied to Ready4Module
}
