% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{enhance}
\alias{enhance}
\title{Enhance a dataset by adding new elements}
\usage{
enhance(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
enhance() is a method that adds new data fields (columns for tabular data, elements for arrays) and values to extendable objects.
}
