% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{enhance}
\alias{enhance}
\title{Enhance an instance of a class with new data, transforming the original object into an instance of a child class}
\usage{
enhance(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
enhance() is a method that enhances an instance of a class with new data, transforming the original object into an instance of a child class.
}
