% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{enhance}
\alias{enhance}
\title{Enhance a ready4 framework module (or sub-module) with new data items}
\usage{
enhance(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
enhance() is a method that adds new data fields and values to extendable objects in a ready4 framework module (or sub-module).
}
