% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{enhance}
\alias{enhance}
\title{Enhance}
\usage{
enhance(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
enhance() is an Enhance function that enhances an instance of a class by adding one or more elements. Specifically, this function implements an algorithm to enhance. The function is called for its side effects and does not return a value.
}
