% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_enhanceSlot.R
\name{enhance}
\alias{enhance}
\alias{enhanceSlot}
\alias{enhanceSlot-Ready4Launch}
\alias{enhanceSlot,Ready4Launch-method}
\title{Enhance an instance of a class by adding new data fields and values to an object stored in a slot of that instance}
\usage{
enhance(x, ...)

enhanceSlot(x, ...)

\S4method{enhanceSlot}{Ready4Launch}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Launch}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
enhanceSlot() is a method that enhances an instance of a class by adding new data fields and values to an object stored in a slot of that instance.

enhanceSlot() is a method that enhances an instance of a class by adding new data fields and values to an object stored in a slot of that instance.

enhanceSlot method applied to Ready4Launch
}
