% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_enhanceSlot.R
\name{enhanceSlot}
\alias{enhanceSlot}
\alias{enhanceSlot-Ready4Module}
\alias{enhanceSlot,Ready4Module-method}
\title{Apply the enhance method to a slot}
\usage{
enhanceSlot(x, slot_nm_1L_chr, ...)

\S4method{enhanceSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
enhanceSlot() is a method that adds new data fields and values to an object stored in a slot of that instance.

enhanceSlot method applied to Ready4Module
}
