% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_enhanceSlot.R
\name{enhanceSlot}
\alias{enhanceSlot}
\alias{enhanceSlot-Ready4Module}
\alias{enhanceSlot,Ready4Module-method}
\title{Apply the enhance method to a model module slot}
\usage{
enhanceSlot(x, slot_nm_1L_chr, ...)

\S4method{enhanceSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method.

A ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method.
}
\description{
enhanceSlot() is a convenience method that applies the enhance method to a specified slot a model module.

enhanceSlot method applied to Ready4Module
}
