% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{exhibit}
\alias{exhibit}
\title{Exhibit features of a dataset by printing them to the R console}
\usage{
exhibit(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
exhibit() is a method that prints salient features of a dataset to console.
}
