% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{exhibit}
\alias{exhibit}
\title{Exhibit (print to console) features of a ready4 framework module (or sub-module)}
\usage{
exhibit(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
exhibit() is a method that prints salient features of a ready4 framework module (or sub-module) to console.
}
