% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{exhibit}
\alias{exhibit}
\title{Exhibit features of model module data by printing them to the R console}
\usage{
exhibit(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance) of the same class as that supplied to the method or no return value (when called for side-effects only).
}
\description{
exhibit() is a method that prints to console selected features of data contained in a model module.
}
