% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_digits_from_text}
\alias{get_digits_from_text}
\title{Get digits from text}
\usage{
get_digits_from_text(text_1L_chr)
}
\arguments{
\item{text_1L_chr}{Text (a character vector of length one)}
}
\value{
Digits (a character vector)
}
\description{
get_digits_from_text() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get digits from text. The function returns Digits (a character vector).
}
\keyword{internal}
