% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_from_lup_obj}
\alias{get_from_lup_obj}
\title{Get from lookup table object}
\usage{
get_from_lup_obj(
  data_lookup_tb,
  match_value_xx,
  match_var_nm_1L_chr,
  target_var_nm_1L_chr,
  evaluate_lgl = TRUE
)
}
\arguments{
\item{data_lookup_tb}{Data lookup (a tibble)}

\item{match_value_xx}{Match value (an output object of multiple potential types)}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one)}

\item{target_var_nm_1L_chr}{Target variable name (a character vector of length one)}

\item{evaluate_lgl}{Evaluate (a logical vector), Default: TRUE}
}
\value{
Return object (an output object of multiple potential types)
}
\description{
get_from_lup_obj() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get from lookup table object. Function argument data_lookup_tb specifies the where to look for the required object. The function returns Return object (an output object of multiple potential types).
}
