% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_generics}
\alias{get_generics}
\title{Get generics}
\usage{
get_generics(
  pkg_nm_1L_chr = "ready4",
  return_1L_chr = "all",
  exclude_mthds_for_chr = NA_character_,
  framework_only_1L_lgl = TRUE
)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}

\item{exclude_mthds_for_chr}{Exclude methods for (a character vector), Default: 'NA'}

\item{framework_only_1L_lgl}{Framework only (a logical vector of length one), Default: TRUE}
}
\value{
Generics (a character vector)
}
\description{
get_generics() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get generics. The function returns Generics (a character vector).
}
\keyword{internal}
