% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_methods}
\alias{get_methods}
\title{Get the methods associated with a ready4 model module}
\usage{
get_methods(pkg_nm_1L_chr = "ready4", cls_nm_1L_chr = "Ready4Module")
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'ready4'}

\item{cls_nm_1L_chr}{Class name (a character vector of length one), Default: 'Ready4Module'}
}
\value{
Methods (a character vector)
}
\description{
get_methods() retrieves the ready4 methods that are available for a specified ready4 model module.
}
\examples{
get_methods()
}
