% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_table_from_loc_file}
\alias{get_table_from_loc_file}
\title{Get table from local file}
\usage{
get_table_from_loc_file(
  path_1L_chr,
  force_numeric_1L_lgl = FALSE,
  force_tb_1L_lgl = FALSE,
  heading_rows_1L_int = 1L
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{force_numeric_1L_lgl}{Force numeric (a logical vector of length one), Default: FALSE}

\item{force_tb_1L_lgl}{Force tibble (a logical vector of length one), Default: FALSE}

\item{heading_rows_1L_int}{Heading rows (an integer vector of length one), Default: 1}
}
\value{
Table (an output object of multiple potential types)
}
\description{
get_table_from_loc_file() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get table from local file. The function returns Table (an output object of multiple potential types).
}
\seealso{
\code{\link[readr:read_delim]{readr::read_csv()}}

\code{\link[readxl:read_excel]{readxl::read_excel()}} and \code{\link[readxl:read_excel]{readxl::read_xlsx()}}
}
\keyword{internal}
