% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ingest}
\alias{ingest}
\title{Ingest data}
\usage{
ingest(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
ingest() is a method that ingests data saved in external files into an object in an active R session.
}
