% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ingest}
\alias{ingest}
\title{Import objects in external file formats into R objects stored in working memory.}
\usage{
ingest(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
ingest() is a method that imports objects in external file formats into R objects stored in working memory.
}
