% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ingest}
\alias{ingest}
\title{Ingest}
\usage{
ingest(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
ingest() is an Ingest function that imports objects in external file formats into R objects stored in working memory. Specifically, this function implements an algorithm to ingest. The function is called for its side effects and does not return a value.
}
