% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{investigate}
\alias{investigate}
\title{Investigate data stored in an instance of a class by executing an algorithm to solve an inverse problem}
\usage{
investigate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
investigate() is a method that investigates data stored in an instance of a class by executing an algorithm to solve an inverse problem.
}
