% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{investigate}
\alias{investigate}
\title{Investigate}
\usage{
investigate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
investigate() is a method that executes an algorithm so solve an inverse problem, often through estimating statistical models.
}
