% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{investigate}
\alias{investigate}
\title{Investigate solutions to an inverse problem}
\usage{
investigate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
investigate() is a method that applies an algorithm to solve an inverse problem (ie, given the data,  identify a statistical model that can generate approximations of that data).
}
