% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{investigate}
\alias{investigate}
\title{Investigate solutions to an inverse problem}
\usage{
investigate(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance).
}
\description{
investigate() is a method that applies an algorithm to data contained in a model module in order to solve an inverse problem (ie, identify a statistical model that can generate approximations of that data).
}
