% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ds_releases_tbl}
\alias{make_ds_releases_tbl}
\title{Make dataset releases table}
\usage{
make_ds_releases_tbl(
  ds_dois_chr,
  format_1L_chr = "\%d-\%b-\%Y",
  key_1L_chr = NULL,
  server_1L_chr = "dataverse.harvard.edu",
  as_kbl_1L_lgl = T,
  ...
)
}
\arguments{
\item{ds_dois_chr}{Dataset digital object identifiers (a character vector)}

\item{format_1L_chr}{Format (a character vector of length one), Default: '\%d-\%b-\%Y'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: 'dataverse.harvard.edu'}

\item{as_kbl_1L_lgl}{As kable (a logical vector of length one), Default: T}

\item{...}{Additional arguments}
}
\value{
Dataset releases (an output object of multiple potential types)
}
\description{
make_ds_releases_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataset releases table. The function returns Dataset releases (an output object of multiple potential types).
}
\examples{
make_ds_releases_tbl("10.7910/DVN/RIQTKK", as_kbl_1L_lgl = FALSE)
}
