% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_framework_pkgs_chr}
\alias{make_framework_pkgs_chr}
\title{Make framework packages character vector}
\usage{
make_framework_pkgs_chr(
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0"
)
}
\arguments{
\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}
}
\value{
Framework packages (a character vector)
}
\description{
make_framework_pkgs_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make framework packages character vector. The function returns Framework packages (a character vector).
}
\keyword{internal}
