% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_imports}
\alias{make_gnrc_imports}
\title{make_gnrc_imports}
\usage{
make_gnrc_imports()
}
\arguments{
\item{fn_args_chr}{Function arguments (a character vector)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_arg_desc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make argument description. The function returns Argument description (a character vector).

FUNCTION_DESCRIPTION
}
\details{
Make argument description

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{GenericFunctions}}
\code{\link[stats]{setNames}}
}
