% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_methods_tb}
\alias{make_methods_tb}
\title{Make methods tibble}
\usage{
make_methods_tb(
  packages_tb = NULL,
  exclude_mthds_for_chr = NA_character_,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  module_pkgs_chr = character(0),
  ns_var_nm_1L_chr = "pt_ns_chr",
  path_1L_chr = character(0),
  return_1L_chr = "all"
)
}
\arguments{
\item{packages_tb}{Packages (a tibble), Default: NULL}

\item{exclude_mthds_for_chr}{Exclude methods for (a character vector), Default: 'NA'}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{module_pkgs_chr}{Module packages (a character vector), Default: character(0)}

\item{ns_var_nm_1L_chr}{Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Methods (a tibble)
}
\description{
make_methods_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make methods tibble. The function returns Methods (a tibble).
}
