% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_methods_tb}
\alias{make_methods_tb}
\title{Make a tabular summary of methods associated with ready model modules}
\usage{
make_methods_tb(
  packages_tb = NULL,
  exclude_mthds_for_chr = NA_character_,
  framework_only_1L_lgl = TRUE,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  module_pkgs_chr = character(0),
  ns_var_nm_1L_chr = "pt_ns_chr",
  path_1L_chr = character(0),
  return_1L_chr = "all"
)
}
\arguments{
\item{packages_tb}{Packages (a tibble), Default: NULL}

\item{exclude_mthds_for_chr}{Exclude methods for (a character vector), Default: 'NA'}

\item{framework_only_1L_lgl}{Framework only (a logical vector of length one), Default: TRUE}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{module_pkgs_chr}{Module packages (a character vector), Default: character(0)}

\item{ns_var_nm_1L_chr}{Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Methods (a tibble)
}
\description{
make_ds_releases_tbl() scrapes metadata from Dataverse datasets for which a valid Digital Object Identifier (DOI) has been supplied to create a table summarising the entire release history of these datasets.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  # Likely to take more than one minute to execute.
  make_methods_tb(gh_repo_1L_chr = "ready4-dev/ready4")
\dontshow{\}) # examplesIf}
}
