% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_modules_tb}
\alias{make_modules_tb}
\title{Make modules tibble}
\usage{
make_modules_tb(
  pkg_extensions_tb = NULL,
  cls_extensions_tb = NULL,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  include_1L_chr = "modules",
  what_chr = "all"
)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble), Default: NULL}

\item{cls_extensions_tb}{Class extensions (a tibble), Default: NULL}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{include_1L_chr}{Include (a character vector of length one), Default: 'modules'}

\item{what_chr}{What (a character vector), Default: 'all'}
}
\value{
Modules (a tibble)
}
\description{
make_modules_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make modules tibble. The function returns Modules (a tibble).
}
