% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_modules_tb}
\alias{make_modules_tb}
\title{Make a tabular summary of ready4 model modules and sub-modules}
\usage{
make_modules_tb(
  pkg_extensions_tb = NULL,
  cls_extensions_tb = NULL,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  module_pkgs_chr = character(0),
  include_1L_chr = "modules",
  ns_var_nm_1L_chr = "pt_ns_chr",
  url_stub_1L_chr = "https://ready4-dev.github.io/",
  what_chr = "all"
)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble), Default: NULL}

\item{cls_extensions_tb}{Class extensions (a tibble), Default: NULL}

\item{gh_repo_1L_chr}{Github repository (a character vector of length one), Default: 'ready4-dev/ready4'}

\item{gh_tag_1L_chr}{Github tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{module_pkgs_chr}{Module packages (a character vector), Default: character(0)}

\item{include_1L_chr}{Include (a character vector of length one), Default: 'modules'}

\item{ns_var_nm_1L_chr}{Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'}

\item{url_stub_1L_chr}{Url stub (a character vector of length one), Default: 'https://ready4-dev.github.io/'}

\item{what_chr}{What (a character vector), Default: 'all'}
}
\value{
Modules (a tibble)
}
\description{
make_methods_tb() scrapes the documentation websites of all libraries of ready4 modules in a specified GitHub organisation and then creates a tabular summary of vignette examples of ready4 module methods.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  # Likely to take more than one minute to execute.
  make_modules_tb(gh_repo_1L_chr = "ready4-dev/ready4")
\dontshow{\}) # examplesIf}
}
