% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{manufacture}
\alias{manufacture}
\title{Manufacture a novel R object using data contained in an instance of a class}
\usage{
manufacture(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
manufacture() is a method that manufactures a novel R object using data contained in an instance of a class.
}
