% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{manufacture}
\alias{manufacture}
\title{Manufacture a new object}
\usage{
manufacture(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
manufacture() is a method that creates a new object (other than a model module or sub-module).
}
