% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{manufacture}
\alias{manufacture}
\title{Manufacture}
\usage{
manufacture(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
manufacture() is a Manufacture function that creates a novel R object. Specifically, this function implements an algorithm to manufacture. The function is called for its side effects and does not return a value.
}
