% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Metamorphose an instance of a class into an instance of a different (non-child) class}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
metamorphose() is a method that metamorphoses an instance of a class into an instance of a different (non-child) class.
}
