% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Transform an instance of a class into an object with different structural properties.}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
metamorphose() is a method that transforms an instance of a class into an object with different structural properties.
}
