% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Metamorphose a ready4 model module to a ready4 model module of a different class}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{A ready4 model module (an instance of a class that inherits from Ready4Module)}

\item{...}{Additional arguments}
}
\value{
A ready4 model module (an instance of a class that inherits from Ready4Module) of a different class to that supplied to the method.
}
\description{
metamorphose() is a method that transforms a ready4 model module into a ready4 model module of a different class.
}
