% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Metamorphose}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
metamorphose() is a Metamorphose function that transforms an instance of a class into an object with different structural properties. Specifically, this function implements an algorithm to metamorphose. The function is called for its side effects and does not return a value.
}
