% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{metamorphose}
\alias{metamorphose}
\title{Metamorphose a model module to a model module of a different (non-inheriting) class}
\usage{
metamorphose(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) of a different class to that supplied to the method.
}
\description{
metamorphose() is a method that transforms a model module into a model module of a different (non-inheriting) class.
}
