% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_metamorphoseSlot.R
\name{metamorphoseSlot}
\alias{metamorphoseSlot}
\alias{metamorphoseSlot-Ready4Launch}
\alias{metamorphoseSlot,Ready4Launch-method}
\title{Metamorphose data stored in a slot of a class instance into an object of a different (non-child) class}
\usage{
metamorphoseSlot(x, ...)

\S4method{metamorphoseSlot}{Ready4Launch}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Launch}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
metamorphoseSlot() is a method that metamorphoses data stored in a slot of a class instance into an object of a different (non-child) class.

metamorphoseSlot method applied to Ready4Launch
}
