% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_metamorphoseSlot.R
\name{metamorphoseSlot}
\alias{metamorphoseSlot}
\alias{metamorphoseSlot-Ready4Module}
\alias{metamorphoseSlot,Ready4Module-method}
\title{Apply the metamorphose method to a slot}
\usage{
metamorphoseSlot(x, slot_nm_1L_chr, ...)

\S4method{metamorphoseSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
metamorphoseSlot() is a method that transforms a ready4 object stored in a slot of an object into a ready4 object of a different class.

metamorphoseSlot method applied to Ready4Module
}
