% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_metamorphoseSlot.R
\name{metamorphoseSlot}
\alias{metamorphoseSlot}
\alias{metamorphoseSlot-Ready4Module}
\alias{metamorphoseSlot,Ready4Module-method}
\title{Apply the metamorphose method to a model module slot}
\usage{
metamorphoseSlot(x, slot_nm_1L_chr, ...)

\S4method{metamorphoseSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module).

A ready4 model module (an instance of a class that inherits from Ready4Module).
}
\description{
metamorphoseSlot() is a convenience method that applies the metamorphose method to a specified slot of a model module.

metamorphoseSlot method applied to Ready4Module
}
