% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_data}
\alias{print_data}
\title{Print data}
\usage{
print_data(
  datasets_tb,
  by_dv_1L_lgl = F,
  root_1L_chr = "https://dataverse.harvard.edu/dataverse/",
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  what_1L_chr = "all",
  ...
)
}
\arguments{
\item{datasets_tb}{Datasets (a tibble)}

\item{by_dv_1L_lgl}{By dataverse (a logical vector of length one), Default: F}

\item{root_1L_chr}{Root (a character vector of length one), Default: 'https://dataverse.harvard.edu/dataverse/'}

\item{scroll_height_1L_chr}{Scroll height (a character vector of length one), Default: character(0)}

\item{scroll_width_1L_chr}{Scroll width (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}

\item{...}{Additional arguments}
}
\value{
Datasets (a kable)
}
\description{
print_data() is a Print function that prints output to console Specifically, this function implements an algorithm to print data. The function is called for its side effects and does not return a value.
}
\examples{
datasets_tb <- get_datasets_tb("ready4-dev/ready4")
print_data(datasets_tb, by_dv_1L_lgl = TRUE)
print_data(datasets_tb, what_1L_chr = "real")
print_data(datasets_tb, what_1L_chr = "fakes")

}
