% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_dss}
\alias{print_dss}
\title{Print datasets}
\usage{
print_dss(
  datasets_tb,
  filter_cdns_ls = NULL,
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  toy_data_dv_1L_chr = "fakes",
  what_1L_chr = "all",
  ...
)
}
\arguments{
\item{datasets_tb}{Datasets (a tibble)}

\item{filter_cdns_ls}{Filter conditions (a list), Default: NULL}

\item{scroll_height_1L_chr}{Scroll height (a character vector of length one), Default: character(0)}

\item{scroll_width_1L_chr}{Scroll width (a character vector of length one), Default: character(0)}

\item{toy_data_dv_1L_chr}{Toy data dataverse (a character vector of length one), Default: 'fakes'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}

\item{...}{Additional arguments}
}
\value{
Datasets (a kable)
}
\description{
print_dss() is a Print function that prints output to console. Specifically, this function implements an algorithm to print datasets. The function is called for its side effects and does not return a value.
}
\keyword{internal}
