% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_dvs}
\alias{print_dvs}
\title{Print dataverses}
\usage{
print_dvs(dvs_tb)
}
\arguments{
\item{dvs_tb}{Dataverses (a tibble)}
}
\value{
dvs_kbl (An object)
}
\description{
print_dvs() is a Print function that prints output to console Specifically, this function implements an algorithm to print dataverses. The function is called for its side effects and does not return a value.
}
\keyword{internal}
