% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_modules}
\alias{print_modules}
\title{Print modules}
\usage{
print_modules(modules_tb, what_1L_chr = "All")
}
\arguments{
\item{modules_tb}{Modules (a tibble)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'All'}
}
\value{
modules_kbl (An object)
}
\description{
print_modules() is a Print function that prints output to console Specifically, this function implements an algorithm to print modules. The function is called for its side effects and does not return a value.
}
