% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_modules}
\alias{print_modules}
\title{Print a table of ready4 model modules}
\usage{
print_modules(
  modules_tb,
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  what_1L_chr = "All",
  ...
)
}
\arguments{
\item{modules_tb}{Modules (a tibble)}

\item{scroll_height_1L_chr}{Scroll height (a character vector of length one), Default: character(0)}

\item{scroll_width_1L_chr}{Scroll width (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'All'}

\item{...}{Additional arguments}
}
\value{
Modules (a kable)
}
\description{
print_modules() formats the output of either get_modules_tb() or make_modules_tb() as HTML.
}
\examples{
modules_tb <- get_modules_tb("ready4-dev/ready4")
# Print sub-modules
print_modules(modules_tb, what_1L_chr = "S3")
# Print full-modules
print_modules(modules_tb, what_1L_chr = "S4")
}
