% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_packages}
\alias{print_packages}
\title{Print packages}
\usage{
print_packages(pkg_extensions_tb = NULL)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble), Default: NULL}
}
\value{
pkg_extensions_kbl (An object)
}
\description{
print_packages() is a Print function that prints output to console Specifically, this function implements an algorithm to print packages. The function is called for its side effects and does not return a value.
}
