% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_vignettes}
\alias{print_vignettes}
\title{Print vignettes}
\usage{
print_vignettes(
  pkg_extensions_tb = NULL,
  include_1L_chr = "modules",
  module_pkgs_chr = character(0),
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  ...
)
}
\arguments{
\item{pkg_extensions_tb}{Package extensions (a tibble), Default: NULL}

\item{include_1L_chr}{Include (a character vector of length one), Default: 'modules'}

\item{module_pkgs_chr}{Module packages (a character vector), Default: character(0)}

\item{scroll_height_1L_chr}{Scroll height (a character vector of length one), Default: character(0)}

\item{scroll_width_1L_chr}{Scroll width (a character vector of length one), Default: character(0)}

\item{...}{Additional arguments}
}
\value{
Vignettes (a kable)
}
\description{
print_vignettes() is a Print function that prints output to console Specifically, this function implements an algorithm to print vignettes. The function is called for its side effects and does not return a value.
}
\keyword{internal}
