% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procure}
\alias{procure}
\title{Procure (get) data}
\usage{
procure(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
procure() is a method that retrieves data contained within an object.
}
