% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procure}
\alias{procure}
\title{Procure data by executing a search and retrieval algorithm using data contained in an instance of a class}
\usage{
procure(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
procure() is a method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class.
}
