% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procure}
\alias{procure}
\title{Procure}
\usage{
procure(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
procure() is a method that searches and retrieves requested data from a specified source.
}
