% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{procure}
\alias{procure}
\title{Procure items from a dataset}
\usage{
procure(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
An object of the same class as that supplied to the method or of one of the same classes that constitute the input object's slots or elements.
}
\description{
procure() is a "getter"  method that retrieves data contained within a model module or sub-module.
}
