% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_procureSlot.R
\name{procureSlot}
\alias{procureSlot}
\alias{procureSlot-Ready4Module}
\alias{procureSlot,Ready4Module-method}
\title{Procure data contained in a slot of a class instance}
\usage{
procureSlot(x, ...)

\S4method{procureSlot}{Ready4Module}(x, slot_nm_1L_chr)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
procureSlot() is a method that procures data contained in a slot of an instance of a class.

procureSlot method applied to Ready4Module
}
