% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_procureSlot.R
\name{procureSlot}
\alias{procureSlot}
\alias{procureSlot-Ready4Module}
\alias{procureSlot,Ready4Module-method}
\title{Procure (get) data from an element of an object}
\usage{
procureSlot(x, slot_nm_1L_chr, ...)

\S4method{procureSlot}{Ready4Module}(x, slot_nm_1L_chr)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
y (an object)
}
\description{
procureSlot() is a "getter" method that retrieves data contained in a slot of an S4 object.

procureSlot method applied to Ready4Module
}
