% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_procureSlot.R
\name{procureSlot}
\alias{procureSlot}
\alias{procureSlot-Ready4Module}
\alias{procureSlot,Ready4Module-method}
\title{Procure (get) data from a slot}
\usage{
procureSlot(x, slot_nm_1L_chr, ...)

\S4method{procureSlot}{Ready4Module}(x, slot_nm_1L_chr, use_procure_mthd_1L_lgl = F, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}

\item{use_procure_mthd_1L_lgl}{Use procure method (a length one logical vector)}
}
\value{
y (an object)
}
\description{
procureSlot() is a "getter" method that, depending on input arguments, retrieves either the value of data contained in a specified model module slot (default behaviour) or the value returned by applying the procure method to the slot.

procureSlot method applied to Ready4Module
}
