% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_procureSlot.R
\name{procureSlot}
\alias{procureSlot}
\alias{procureSlot-Ready4Module}
\alias{procureSlot,Ready4Module-method}
\title{Procure (get) data from a slot}
\usage{
procureSlot(x, slot_nm_1L_chr, ...)

\S4method{procureSlot}{Ready4Module}(x, slot_nm_1L_chr, use_procure_mthd_1L_lgl = FALSE, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}

\item{use_procure_mthd_1L_lgl}{Use procure method (a length one logical vector)}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or an instance of a class contained in that Ready4Module's slots.

A ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or an instance of a class contained in that Ready4Module's slots.
}
\description{
procureSlot() is a "getter" method that, depending on input arguments, retrieves either data contained in a specified model module slot (the default behaviour) or the value returned by applying the procure method to the specified slot.

procureSlot method applied to Ready4Module
}
\examples{
X <- Ready4Module()
procureSlot(X, "dissemination_1L_chr")
}
