% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{prognosticate}
\alias{prognosticate}
\title{Prognosticate by applying an algorithm to solve a forward problem through simulation or prediction}
\usage{
prognosticate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
prognosticate() is a method that prognosticates by applying an algorithm to solve a forward problem through simulation or prediction.
}
