% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{prognosticate}
\alias{prognosticate}
\title{Prognosticate}
\usage{
prognosticate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
prognosticate() is a method that executes an algorithm for solving forward problems through simulation or prediction.
}
