% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{prognosticate}
\alias{prognosticate}
\title{Prognosticate (make predictions) by solving a forward problem}
\usage{
prognosticate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
prognosticate() is a method that uses statistical or simulation models to solve a forward problem (i.e., given a model, calculate what is expected to be observed).
}
