% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{prognosticate}
\alias{prognosticate}
\title{Prognosticate}
\usage{
prognosticate(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
prognosticate() is a Prognosticate function that executes an algorithm for solving forward problems through simulation or prediction. Specifically, this function implements an algorithm to prognosticate. The function is called for its side effects and does not return a value.
}
