% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_prognosticateSlot.R
\name{prognosticateSlot}
\alias{prognosticateSlot}
\alias{prognosticateSlot-Ready4Module}
\alias{prognosticateSlot,Ready4Module-method}
\title{Apply the prognosticate method to a model module slot}
\usage{
prognosticateSlot(x, slot_nm_1L_chr, ...)

\S4method{prognosticateSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module).

A ready4 model module (an instance of a class that inherits from Ready4Module).
}
\description{
prognosticateSlot() is a convenience method that applies the prognosticate method to a specified slot of a model module.

prognosticateSlot method applied to Ready4Module
}
