% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ratify}
\alias{ratify}
\title{Ratify that a dataset meets validity criteria}
\usage{
ratify(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
ratify() is a method that validates that a model dataset conforms to specified internal consistency criteria, potentially modifying the dataset so that these criteria are met.
}
