% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ratify}
\alias{ratify}
\title{Ratifie that an instance of a class conforms to specified criteria}
\usage{
ratify(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
ratify() is a method that ratifies that an instance of a class conforms to specified criteria.
}
