% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ratify}
\alias{ratify}
\title{Ratify (validate) a ready4 framework module / sub-module}
\usage{
ratify(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
ratify() is a method that validates that a ready4 framework module / sub-module conforms to specified criteria, potentially modifying the object so that these criteria are met.
}
