% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{ratifySlot}
\alias{ratifySlot}
\title{Ratifie that data contained in a slot of a class instance conforms to specified criteria}
\usage{
ratifySlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
ratifySlot() is a method that ratifies that data contained in a slot of a class instance conforms to specified criteria.
}
