% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_ratifySlot.R
\name{ratifySlot}
\alias{ratifySlot}
\alias{ratifySlot-Ready4Launch}
\alias{ratifySlot,Ready4Launch-method}
\title{Ratifie that data contained in a slot of a class instance conforms to specified criteria}
\usage{
ratifySlot(x, ...)

\S4method{ratifySlot}{Ready4Launch}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Launch}

\item{...}{Additional arguments}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
ratifySlot() is a method that ratifies that data contained in a slot of a class instance conforms to specified criteria.

ratifySlot method applied to Ready4Launch
}
