% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4.R, R/ready4-package.R
\docType{package}
\name{ready4-package}
\alias{ready4}
\alias{ready4-package}
\title{ready4: Implement Transparent, Reusable and Updatable Computational Health Economic Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Foundational elements (classes, generics, methods, functions) of a prototype software framework for developing modular, open source computational health economic models. The foundation provided by 'ready4' is extended by other R packages. For detailed documentation visit \url{https://www.ready4-dev.com/}. For a scientific summary read \href{https://arxiv.org/abs/2310.14138}{arXiv:2310.14138}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Foundational elements (classes, generics, methods, functions) of a prototype software framework for developing modular, open source computational health economic models. The foundation provided by 'ready4' is extended by other R packages. For detailed documentation visit \url{https://www.ready4-dev.com/}. For a scientific summary read \href{https://arxiv.org/abs/2310.14138}{arXiv:2310.14138}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
