% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4.R, R/ready4-package.R
\docType{package}
\name{ready4-package}
\alias{ready4}
\alias{ready4-package}
\title{ready4: Implement a Modular, Open Source Computational Model of Youth Mental Health Systems}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4 provides bare bones foundational elements (classes, generics, methods, functions) of a modular, open source model of the systems shaping young people's mental health. The base elements in this package are extended by other R packages - see https://www.ready4-dev.com/ for details. This development version of the ready4 package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the author (matthew.hamilton@orygen.org.au).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4 provides bare bones foundational elements (classes, generics, methods, functions) of a modular, open source model of the systems shaping young people's mental health. The base elements in this package are extended by other R packages - see https://www.ready4-dev.com/ for details. This development version of the ready4 package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the author (matthew.hamilton@orygen.org.au).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
