% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4.R, R/ready4-package.R
\docType{package}
\name{ready4-package}
\alias{ready4-package}
\title{ready4: Implement Modular and Open-Source Health Economic Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programming syntax, a template model module and tools to help maintain a modular and open-source health economic model's project documentation website. These elements are the foundation for a prototype software framework to support replicable and transferable health economic models. The software framework is extended by other R libraries. For detailed documentation about the framework and how to use it visit \url{https://www.ready4-dev.com/}. For a background to the methodological issues that the framework is attempting to help solve, see Hamilton et al. (2024) \doi{10.48550/arXiv.2310.14138}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programming syntax, a template model module and tools to help maintain a modular and open-source health economic model's project documentation website. These elements are the foundation for a prototype software framework to support replicable and transferable health economic models. The software framework is extended by other R libraries. For detailed documentation about the framework and how to use it visit \url{https://www.ready4-dev.com/}. For a background to the methodological issues that the framework is attempting to help solve, see Hamilton et al. (2024) \doi{10.48550/arXiv.2310.14138}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4/}
  \item \url{https://github.com/ready4-dev/ready4}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
