% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Private.R
\docType{class}
\name{Ready4Private-class}
\alias{Ready4Private-class}
\alias{Ready4Private}
\title{Ready4Private}
\description{
A virtual class denoting a module of the ready4 representational system that contains data not intended for public dissemination.
}
