% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4Public.R
\docType{class}
\name{Ready4Public-class}
\alias{Ready4Public-class}
\alias{Ready4Public}
\title{Ready4Public}
\description{
A virtual class denoting a module of the ready4 representational system hat is suitable for public dissemination in its current form.
}
\section{Slots}{

\describe{
\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

