% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{reckon}
\alias{reckon}
\title{Reckon (calculate) a value}
\usage{
reckon(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
reckon() is a method that  performs a calculation using data contained in a ready4 framework module / sub-module.
}
