% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{reckon}
\alias{reckon}
\title{Perform a calculation.}
\usage{
reckon(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
reckon() is a method that performs a calculation.
}
