% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{reckon}
\alias{reckon}
\title{Reckon}
\usage{
reckon(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
reckon() is a Reckon function that performs a calculation. Specifically, this function implements an algorithm to reckon. The function is called for its side effects and does not return a value.
}
