% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{reckon}
\alias{reckon}
\title{Reckon (calculate) a value}
\usage{
reckon(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A numeric class.
}
\description{
reckon() is a method that performs a calculation using data contained in a model module (or sub-module).
}
