% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_reckonSlot.R
\name{reckonSlot}
\alias{reckonSlot}
\alias{reckonSlot-Ready4Module}
\alias{reckonSlot,Ready4Module-method}
\title{Apply the reckon method to a model module slot}
\usage{
reckonSlot(x, slot_nm_1L_chr, ...)

\S4method{reckonSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
A numeric class.

A numeric class.
}
\description{
reckonSlot() is a convenience method that applies the reckon method to a specified slot of a model module.

reckonSlot method applied to Ready4Module
}
