% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{renew}
\alias{renew}
\title{Renew (update) a ready4 framework module / sub-module}
\usage{
renew(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
renew() is a method that updates a ready4 framework module / sub-module with new values.
}
