% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{renew}
\alias{renew}
\title{Renew values in a dataset}
\usage{
renew(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
renew() is a method that updates a dataset contained within a model module or sub-module with new values. A "setter" method.
}
