% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{renew}
\alias{renew}
\title{Renew}
\usage{
renew(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
renew() is a Renew function that updates an instance of a class with new values. Specifically, this function implements an algorithm to renew. The function is called for its side effects and does not return a value.
}
