% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_renewSlot.R
\name{renewSlot}
\alias{renewSlot}
\alias{renewSlot-Ready4Launch}
\alias{renewSlot,Ready4Launch-method}
\title{Renew a slot of a class instance by updating it with new data}
\usage{
renewSlot(x, ...)

\S4method{renewSlot}{Ready4Launch}(x, new_val_xx, slot_nm_1L_chr)
}
\arguments{
\item{x}{An object of class Ready4Launch}

\item{...}{Additional arguments}

\item{new_val_xx}{New value (slot dependent object type)}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}
}
\description{
renewSlot() is a method that renews a slot of a class instance by updating it with new data.

renewSlot method applied to Ready4Launch
}
