% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{renewSlot}
\alias{renewSlot}
\title{Update a specified slot of a class instance with new values.}
\usage{
renewSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
renewSlot() is a method that updates a specified slot of a class instance with new values.
}
