% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
report() is a Report function that authors a report. Specifically, this function implements an algorithm to report. The function is called for its side effects and does not return a value.
}
