% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{share}
\alias{share}
\title{Share data contained in an instance of a class via an online repository}
\usage{
share(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
share() is a method that shares data contained in an instance of a class via an online repository.
}
