% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{share}
\alias{share}
\title{Process output for public disemination and upload to an online repository along with required metadata.}
\usage{
share(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
share() is a method that processes output for public disemination and uploads to an online repository along with required metadata.
}
