% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{share}
\alias{share}
\title{Share}
\usage{
share(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
share() is a Share function that processes output for public disemination and uploads to an online repository along with required metadata. Specifically, this function implements an algorithm to share. The function is called for its side effects and does not return a value.
}
