% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{share}
\alias{share}
\title{Share data via an online repository}
\usage{
share(x, ...)
}
\arguments{
\item{x}{A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance)}

\item{...}{Additional arguments}
}
\value{
A model module (an instance of a class that inherits from Ready4Module) or submodule (any S3 class instance) of the same class as that supplied to the method or no return value (when called for side-effects only).
}
\description{
share() is a method that uploads data contained in a model module to an online repository. If requested, the method will also publish the updated repository.
}
