% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{shareSlot}
\alias{shareSlot}
\title{Share data contained in a slot of a class instance via an online repository}
\usage{
shareSlot(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
shareSlot() is a method that shares data contained in a slot of a class instance via an online repository.
}
