% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_shareSlot.R
\name{shareSlot}
\alias{shareSlot}
\alias{shareSlot-Ready4Module}
\alias{shareSlot,Ready4Module-method}
\title{Apply the share method to a slot}
\usage{
shareSlot(x, slot_nm_1L_chr, ...)

\S4method{shareSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\description{
shareSlot() is a method that uploads data contained in a slot of an S4 object to an online repository. If requested, the method will also publish the updated repository.

shareSlot method applied to Ready4Module
}
