% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_shareSlot.R
\name{shareSlot}
\alias{shareSlot}
\alias{shareSlot-Ready4Module}
\alias{shareSlot,Ready4Module-method}
\title{Apply the share method to a model module slot}
\usage{
shareSlot(x, slot_nm_1L_chr, ...)

\S4method{shareSlot}{Ready4Module}(x, slot_nm_1L_chr, ...)
}
\arguments{
\item{x}{An object of class Ready4Module}

\item{slot_nm_1L_chr}{Slot name (a length one character vector)}

\item{...}{Additional arguments}
}
\value{
Either a model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called purely for side effects).

Either a ready4 model module (an instance of a class that inherits from Ready4Module) of the same class as that supplied to the method or no value (when called purely for side effects).
}
\description{
shareSlot() is a convenience method that applies the share method to a specified slot of a model module.

shareSlot method applied to Ready4Module
}
