% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_tb_r3}
\alias{update_tb_r3}
\title{Update tibble ready4 submodule}
\usage{
update_tb_r3(
  tb_r3,
  case_when_false_1L_chr = NA_character_,
  case_when_true_1L_chr = NA_character_,
  case_when_true_ls = NULL,
  case_when_var_1L_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  fn = NULL,
  fn_env_ls = NULL,
  slice_indcs_int = NA_integer_,
  tf_false_val_1L_lgl = FALSE
)
}
\arguments{
\item{tb_r3}{Tibble ready4 submodule (a ready4 submodule extension of tibble)}

\item{case_when_false_1L_chr}{Case when false (a character vector of length one), Default: 'NA'}

\item{case_when_true_1L_chr}{Case when true (a character vector of length one), Default: 'NA'}

\item{case_when_true_ls}{Case when true (a list), Default: NULL}

\item{case_when_var_1L_chr}{Case when variable (a character vector of length one), Default: 'NA'}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{fn}{Function (a function), Default: NULL}

\item{fn_env_ls}{Function (a list of environments), Default: NULL}

\item{slice_indcs_int}{Slice indices (an integer vector), Default: NA}

\item{tf_false_val_1L_lgl}{Transform false value (a logical vector of length one), Default: FALSE}
}
\value{
Tibble ready4 submodule (a ready4 submodule extension of tibble)
}
\description{
update_tb_r3() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update tibble ready4 submodule. The function returns Tibble ready4 submodule (a ready4 submodule extension of tibble).
}
\keyword{internal}
