% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_all_tbs_in_tbs_r4_to_csvs}
\alias{write_all_tbs_in_tbs_r4_to_csvs}
\title{Write all tibbles in tibbles ready4 module to comma separated variables files}
\usage{
write_all_tbs_in_tbs_r4_to_csvs(
  tbs_r4,
  r4_name_1L_chr,
  lup_dir_1L_chr,
  pfx_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{tbs_r4}{Tibbles (a ready4 module)}

\item{r4_name_1L_chr}{Ready4 module name (a character vector of length one)}

\item{lup_dir_1L_chr}{Lookup table directory (a character vector of length one)}

\item{pfx_1L_chr}{Prefix (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
No return value, called for side effects.
}
\description{
write_all_tbs_in_tbs_r4_to_csvs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write all tibbles in tibbles ready4 module to comma separated variables files. The function is called for its side effects and does not return a value.
}
\keyword{internal}
