% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_citation_cff}
\alias{write_citation_cff}
\title{Write citation cff}
\usage{
write_citation_cff(
  pkg_desc_ls,
  citation_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  publisher_1L_chr = "Zenodo"
)
}
\arguments{
\item{pkg_desc_ls}{Package description (a list)}

\item{citation_chr}{Citation (a character vector)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{publisher_1L_chr}{Publisher (a character vector of length one), Default: 'Zenodo'}
}
\value{
No return value, called for side effects.
}
\description{
write_citation_cff() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write citation cff. The function is called for its side effects and does not return a value.
}
\keyword{internal}
