% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_env_objs_to_dv}
\alias{write_env_objs_to_dv}
\title{Write environment objects to dataverse}
\usage{
write_env_objs_to_dv(
  env_objects_ls,
  descriptions_chr,
  ds_url_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  options_chr = c("Y", "N"),
  publish_dv_1L_lgl = FALSE,
  piggyback_desc_1L_chr = "Documentation",
  piggyback_tag_1L_chr = "Documentation_0.0",
  piggyback_to_1L_chr = character(0),
  prerelease_1L_lgl = TRUE,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{env_objects_ls}{Environment objects (a list)}

\item{descriptions_chr}{Descriptions (a character vector)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{publish_dv_1L_lgl}{Publish dataverse (a logical vector of length one), Default: FALSE}

\item{piggyback_desc_1L_chr}{Piggyback description (a character vector of length one), Default: 'Documentation'}

\item{piggyback_tag_1L_chr}{Piggyback tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{piggyback_to_1L_chr}{Piggyback to (a character vector of length one), Default: character(0)}

\item{prerelease_1L_lgl}{Prerelease (a logical vector of length one), Default: TRUE}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
File identities (an integer vector)
}
\description{
write_env_objs_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write environment objects to dataverse. The function returns File identities (an integer vector).
}
\keyword{internal}
