% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_from_dv}
\alias{write_fls_from_dv}
\title{Write files from dataverse}
\usage{
write_fls_from_dv(
  files_tb,
  fl_ids_int,
  ds_url_1L_chr,
  local_dv_dir_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  options_chr = c("Y", "N"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{files_tb}{Files (a tibble)}

\item{fl_ids_int}{File identities (an integer vector)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{local_dv_dir_1L_chr}{Local dataverse directory (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
No return value, called for side effects.
}
\description{
write_fls_from_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files from dataverse. The function is called for its side effects and does not return a value.
}
\keyword{internal}
