% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_to_repo}
\alias{write_fls_to_repo}
\title{Write files to repository}
\usage{
write_fls_to_repo(
  paths_chr,
  descriptions_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  ds_url_1L_chr = character(0),
  ds_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  options_chr = c("Y", "N"),
  piggyback_desc_1L_chr = "Documentation",
  piggyback_tag_1L_chr = "Documentation_0.0",
  piggyback_to_1L_chr = character(0),
  prerelease_1L_lgl = TRUE,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{paths_chr}{Paths (a character vector)}

\item{descriptions_chr}{Descriptions (a character vector)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one), Default: character(0)}

\item{ds_ls}{Dataset (a list), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{piggyback_desc_1L_chr}{Piggyback description (a character vector of length one), Default: 'Documentation'}

\item{piggyback_tag_1L_chr}{Piggyback tag (a character vector of length one), Default: 'Documentation_0.0'}

\item{piggyback_to_1L_chr}{Piggyback to (a character vector of length one), Default: character(0)}

\item{prerelease_1L_lgl}{Prerelease (a logical vector of length one), Default: TRUE}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Identities (an integer vector)
}
\description{
write_fls_to_repo() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files to repository. The function returns Identities (an integer vector).
}
\keyword{internal}
