% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_from_tmp}
\alias{write_from_tmp}
\title{Write from temporary}
\usage{
write_from_tmp(
  tmp_paths_chr,
  dest_paths_chr,
  args_ls_ls = list(NULL),
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  edit_fn_ls = list(NULL),
  options_chr = c("Y", "N")
)
}
\arguments{
\item{tmp_paths_chr}{Temporary paths (a character vector)}

\item{dest_paths_chr}{Destination paths (a character vector)}

\item{args_ls_ls}{Arguments (a list of lists), Default: list(NULL)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{edit_fn_ls}{Edit (a list of functions), Default: list(NULL)}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
No return value, called for side effects.
}
\description{
write_from_tmp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write from temporary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
