% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_tb_to_csv}
\alias{write_tb_to_csv}
\title{Write tibble to comma separated variables file}
\usage{
write_tb_to_csv(
  tbs_r4,
  slot_nm_1L_chr,
  r4_name_1L_chr,
  lup_dir_1L_chr,
  pfx_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{tbs_r4}{Tibbles (a ready4 module)}

\item{slot_nm_1L_chr}{Slot name (a character vector of length one)}

\item{r4_name_1L_chr}{Ready4 module name (a character vector of length one)}

\item{lup_dir_1L_chr}{Lookup table directory (a character vector of length one)}

\item{pfx_1L_chr}{Prefix (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
No return value, called for side effects.
}
\description{
write_tb_to_csv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write tibble to comma separated variables file. The function is called for its side effects and does not return a value.
}
\keyword{internal}
