% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_dv_with_wait}
\alias{write_to_dv_with_wait}
\title{Write to dataverse with wait}
\usage{
write_to_dv_with_wait(
  dss_tb,
  dv_nm_1L_chr,
  ds_url_1L_chr,
  parent_dv_dir_1L_chr,
  paths_to_dirs_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  make_local_copy_1L_lgl = FALSE,
  options_chr = c("Y", "N"),
  paths_are_rltv_1L_lgl = TRUE,
  inc_fl_types_chr = NA_character_,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  wait_time_in_secs_int = 5L
)
}
\arguments{
\item{dss_tb}{Datasets (a tibble)}

\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{parent_dv_dir_1L_chr}{Parent dataverse directory (a character vector of length one)}

\item{paths_to_dirs_chr}{Paths to directories (a character vector)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{make_local_copy_1L_lgl}{Make local copy (a logical vector of length one), Default: FALSE}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{paths_are_rltv_1L_lgl}{Paths are relative (a logical vector of length one), Default: TRUE}

\item{inc_fl_types_chr}{Include file types (a character vector), Default: 'NA'}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{wait_time_in_secs_int}{Wait time in secs (an integer vector), Default: 5}
}
\value{
Dataset (a list)
}
\description{
write_to_dv_with_wait() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to dataverse with wait. The function returns Dataset (a list).
}
\keyword{internal}
