% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_publish_dv_ds}
\alias{write_to_publish_dv_ds}
\title{Write to publish dataverse dataset}
\usage{
write_to_publish_dv_ds(
  dv_ds_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  minor_1L_lgl = FALSE,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{dv_ds_1L_chr}{Dataverse dataset (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{minor_1L_lgl}{Minor (a logical vector of length one), Default: FALSE}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
No return value, called for side effects.
}
\description{
write_to_publish_dv_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to publish dataverse dataset. The function is called for its side effects and does not return a value.
}
\keyword{internal}
