% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_render_post}
\alias{write_to_render_post}
\title{Write to render post}
\usage{
write_to_render_post(
  included_dirs_chr,
  path_to_main_dir_1L_chr,
  is_rmd_1L_lgl = T
)
}
\arguments{
\item{included_dirs_chr}{Included directories (a character vector)}

\item{path_to_main_dir_1L_chr}{Path to main directory (a character vector of length one)}

\item{is_rmd_1L_lgl}{Is Markdown (a logical vector of length one), Default: T}
}
\description{
write_to_render_post() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to render post. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
