% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ws}
\alias{write_ws}
\title{Write ready4 software development local directories}
\usage{
write_ws(
  path_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
No return value, called for side effects.
}
\description{
write_ws() creates a standardised directory structure as a local development environment for modelling projects developed with the ready4 framework.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  write_ws(tempdir())
\dontshow{\}) # examplesIf}
}
